% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuf.R
\name{cbind.all}
\alias{cbind.all}
\title{Combine dataframes by Columns}
\usage{
\method{cbind}{all}(...)
}
\arguments{
\item{...}{dataframes to be combined by columns}
}
\value{
A matrix combining the \code{...} arguments column-wise
}
\description{
Allows cbind to work with dataframes with different columns
by filling missing values with NA
found on https://stackoverflow.com/questions/7962267/
}
\examples{
a <- data.frame()
b <- data.frame(a = 1, b = 2, c = 3)
df <- cbind.all(a, b)
}
