% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysis.R
\name{CGMTPEa}
\alias{CGMTPEa}
\title{#' R6 Class Representing a CGMTPE analysis}
\description{
TODO: change name
The CGMTPE Analysis object "analysis" is an environment containing all
objects used for estimation and TPE analysis
(varieties, environments, grids, maps, etc.)
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the CGMTPE analysis}

\item{\code{model}}{A character string with the name of the crop model
used for simulations (only "Samara" is supported for now)}

\item{\code{varieties}}{A list of varieties objects}

\item{\code{TPEanalysis}}{A list of TPE analysis objects}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CGMTPEa-new}{\code{CGMTPEa$new()}}
\item \href{#method-CGMTPEa-initMessage}{\code{CGMTPEa$initMessage()}}
\item \href{#method-CGMTPEa-set_name}{\code{CGMTPEa$set_name()}}
\item \href{#method-CGMTPEa-set_model}{\code{CGMTPEa$set_model()}}
\item \href{#method-CGMTPEa-set_var}{\code{CGMTPEa$set_var()}}
\item \href{#method-CGMTPEa-get_varNames}{\code{CGMTPEa$get_varNames()}}
\item \href{#method-CGMTPEa-get_genotypes}{\code{CGMTPEa$get_genotypes()}}
\item \href{#method-CGMTPEa-get_varid}{\code{CGMTPEa$get_varid()}}
\item \href{#method-CGMTPEa-createTPE}{\code{CGMTPEa$createTPE()}}
\item \href{#method-CGMTPEa-createGrid}{\code{CGMTPEa$createGrid()}}
\item \href{#method-CGMTPEa-genClimate}{\code{CGMTPEa$genClimate()}}
\item \href{#method-CGMTPEa-runGridSim}{\code{CGMTPEa$runGridSim()}}
\item \href{#method-CGMTPEa-createMap}{\code{CGMTPEa$createMap()}}
\item \href{#method-CGMTPEa-runClustering}{\code{CGMTPEa$runClustering()}}
\item \href{#method-CGMTPEa-plotMap}{\code{CGMTPEa$plotMap()}}
\item \href{#method-CGMTPEa-runEstimation}{\code{CGMTPEa$runEstimation()}}
\item \href{#method-CGMTPEa-clone}{\code{CGMTPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-new"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$new(
  name = "CGMTPE_analysis",
  model = "Samara",
  varieties = NA,
  genotypes = NA,
  vparameters = NA,
  environments = NA,
  eparameters = NA,
  eName = "estim1",
  weathers = NA,
  observations = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}

\item{\code{model}}{A character string with the name of the crop model used
for simulations}

\item{\code{varieties}}{A list of varieties names}

\item{\code{genotypes}}{Optional. A list of alternate variety names}

\item{\code{vparameters}}{A vector or dataframe with variety specific parameters}

\item{\code{environments}}{A list of environments names}

\item{\code{eparameters}}{A vector or dataframe with environment specific
parameters}

\item{\code{eName}}{A character string identifier of the initial estimation
object}

\item{\code{weathers}}{A list (each entry is an environment)
of dataframes with weather data}

\item{\code{observations}}{A list (each entry is a variety) of lists
(each entry is an environment) of dataframes with observations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CGMTPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$initMessage(vParamMissings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vParamMissings}}{Name of varieties without parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPE analysis name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPE analysis name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set varieties, this is a list of TPEvar objects,
use with caution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New varieties}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-get_varNames"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-get_varNames}{}}}
\subsection{Method \code{get_varNames()}}{
Get names of all varieties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$get_varNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-get_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-get_genotypes}{}}}
\subsection{Method \code{get_genotypes()}}{
Get names of all genotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$get_genotypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-get_varid"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-get_varid}{}}}
\subsection{Method \code{get_varid()}}{
Get var id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$get_varid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either var name or a var id (will return the var id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-createTPE"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-createTPE}{}}}
\subsection{Method \code{createTPE()}}{
Create a TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$createTPE(name = "TPEa_1")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string of TPE analysis name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$createGrid(
  tpeID = 1,
  name = "g1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{name}}{A character string identifier of the grid}

\item{\code{varID}}{A variety identifier to use for simulation on the grid}

\item{\code{latres}}{A numeric value of the latitude resolution of the grid}

\item{\code{lonres}}{A numeric value of the longitude resolution of the grid}

\item{\code{cols}}{A numeric value of the number of columns in the grid}

\item{\code{rows}}{A numeric value of the number of rows in the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the grid in
decimal degrees}

\item{\code{lat}}{A numeric value of the starting latitude of the grid in
decimal degrees}

\item{\code{multigrid}}{A boolean indicating if one grid should be created
for each genotype with the same cols, rows, lon and lat. If true, varID
will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$genClimate(
  tpeID = 1,
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  seed = NA,
  filesE = F,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{seed}}{A numeric value of the seed to use to generate climate}

\item{\code{filesE}}{Boolean. If weather files already exist}

\item{\code{verbose}}{Boolean. If messages about starting climate generation
should be shown}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$runGridSim(
  tpeID = 1,
  gridID = NA,
  varID = NA,
  soilData = soil,
  latlonData = lat_lon,
  cumulP = cumul,
  traitList = NULL,
  savePath = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A variety identifier to use for
simulation on the grid. If NA, will use the variety attached to grid}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the max for
each year}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create raster map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$createMap(tpeID = 1, name = "map1", bounds = NA, res = 150)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{name}}{A character string defining the name of the map}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map}

\item{\code{res}}{Not used at the moment. A numeric value in sec of the
resolution of world map to use. Options are c(1, 150, 900)
for respectively 30sec, 2.5min, 15min}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-runClustering"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-runClustering}{}}}
\subsection{Method \code{runClustering()}}{
Run TPE clustering by performing Principle Component
Analysis (PCA) and Hierarchical Clustering on Principle Component (HCPC)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$runClustering(
  tpeID = 1,
  mapID = 1,
  traitList = c("GrainYieldPop"),
  nbDim = 5,
  nbClust = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{mapID}}{A value or list of values of map identifiers
(either index or names)on which to perform clustering}

\item{\code{traitList}}{Vector of variable names to be used for PCA}

\item{\code{nbDim}}{Integer of number of dimensions to use for PCA}

\item{\code{nbClust}}{Integer of number of clusters to use for HCPC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$plotMap(tpeID = 1, mapID = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE analysis identifier on which to create grid}

\item{\code{mapID}}{Id of map to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation


Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$runEstimation(
  varID = 1,
  estimID = 1,
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varID}}{A value or list of values of variety identifier
(either index or name)}

\item{\code{estimID}}{A value of estimation identifier (either index or name)}

\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment}

\item{\code{bounds}}{A matrix with lower (row1) and upper (row2) bounds for
each of the parameters in `paramnames` (cols)}

\item{\code{...}}{Additional parameters to be passed to DEoptim.control}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGMTPEa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CGMTPEa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGMTPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
