% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimation.R
\name{check_results}
\alias{check_results}
\title{Run and plot simulation with estimated parameters}
\usage{
check_results(param, weather, estim = NA, paramnames = NA)
}
\arguments{
\item{param}{Dataframe of all parameters used in simulation
(variety and environment)}

\item{weather}{Dataframe with weather data}

\item{estim}{Optionnal, vector with value of estimated parameters}

\item{paramnames}{Optionnal, vector with names of estimated parameters}
}
\description{
TODO
}
\examples{
check_results(param, weather, estim ,paramnames)
}
