% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_tpe.R
\name{create_tpe}
\alias{create_tpe}
\title{Create a TPE analysis object}
\usage{
create_tpe(name = "tpea1", model = "Samara")
}
\arguments{
\item{name}{test}

\item{model}{test}
}
\value{
A R6 class TPE analysis object
}
\description{
Create a TPE analysis object (from R6 class TPEa)
}
\examples{
TPE_analysos <- create_tpe("tpea1","Samara")
}
