% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_estimation.R
\name{estim_param}
\alias{estim_param}
\title{Generic function for parameter estimation over 1 to n environments}
\usage{
estim_param(
  p,
  environments,
  param,
  paramnames,
  weather,
  obs,
  score_fn,
  metric = "RMSE",
  weigh_fn = NA
)
}
\description{
TODO
}
