% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_estim.R
\name{estimP}
\alias{estimP}
\title{#' R6 Class Representing a parameter estimation process}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of estimation}

\item{\code{parent}}{Parent variety}

\item{\code{environments}}{List of environments to use for estimation}

\item{\code{parameters}}{A dataframe with all parameters (variety x environment)
used for simulation, each line is one environment}

\item{\code{DEparams}}{A DEoptim.control object with parameters used for DEoptim
parameter estimation}

\item{\code{paramNames}}{A vector of estimated parameter names}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-estimP-new}{\code{estimP$new()}}
\item \href{#method-estimP-createEnv}{\code{estimP$createEnv()}}
\item \href{#method-estimP-runEstimation}{\code{estimP$runEstimation()}}
\item \href{#method-estimP-clone}{\code{estimP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-new"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE soil object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$new(
  name = "estim1",
  parent,
  environments,
  eparam,
  weathers,
  observations
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of estimation}

\item{\code{parent}}{Parent variety}

\item{\code{environments}}{List of environment names}

\item{\code{eparam}}{Parameters of environment}

\item{\code{weathers}}{Weather of environment}

\item{\code{observations}}{Observations of environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEsoil` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-createEnv"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-createEnv}{}}}
\subsection{Method \code{createEnv()}}{
Create environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$createEnv(name, eparam, weather, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of environment}

\item{\code{eparam}}{Parameters of environment}

\item{\code{weather}}{Weather of environment}

\item{\code{observations}}{Observations of environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation


Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$runEstimation(
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment}

\item{\code{bounds}}{A matrix with lower (col1) and upper (col2) bounds for
each of the parameters in `paramnames` (rows)}

\item{\code{args}}{Additional parameters to be passed to DEoptim.control}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
