% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_tpe.R
\name{et0}
\alias{et0}
\title{Compute et0}
\usage{
et0(tmin, tmax, srad)
}
\arguments{
\item{tmin}{test}

\item{tmax}{test}

\item{srad}{test}
}
\value{
A vector with et0 values
}
\description{
Compute evapotranspiration et0
}
\examples{
et0Values <- et0(tmin,tmax,srad)
}
