% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_estimation.R
\name{fitness_score}
\alias{fitness_score}
\title{Fitness function for parameter estimation of Samara}
\usage{
fitness_score(
  p,
  param,
  paramnames,
  weather,
  obs,
  score_fn,
  idx,
  metric = "RMSE",
  weigh_fn = NA
)
}
\arguments{
\item{p}{A vector of a candidate parameter set}

\item{param}{A dataframe with all Samara parameters}

\item{paramnames}{A vector containing estimated parameter names}

\item{weather}{A dataframe with all weather data}

\item{obs}{A dataframe or list of dataframes with all observation data}

\item{score_fn}{A user-defined function to compute the error between}

\item{idx}{An integer used as the simulation identifier}
}
\value{
The fitness score for the candidate parameter set
}
\description{
Computes the fitness score of a given parameter set candidate
and according to a user-defined function (see examples)
}
\examples{

}
