% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_tpe.R
\name{generateClimate}
\alias{generateClimate}
\title{Generate weather file for Samara from Marksim}
\usage{
generateClimate(lon, lat, rcp, year, yearNb, modelNb, path, pathCLI = NA)
}
\arguments{
\item{variety}{test}
}
\value{
test
}
\description{
test
}
\examples{
test
}
