% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_estimation.R
\name{get_score}
\alias{get_score}
\title{Compute fitness score between observations and simulations}
\usage{
get_score(obs, sim, metric = NA, weight_fn = NA)
}
\arguments{
\item{obs}{Dataframe of observations}

\item{sim}{Dataframe of simulations}

\item{metric}{Character string of the metric to use as computation of the
score (choices are : MAE, RMSE, MSE, rMAE, rRMSE)}

\item{weight_fn}{Optionnal. TODO}
}
\description{
TODO
}
\examples{
get_score(obs, sim, metric, weight_fn)
}
