% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gridpoint.R
\name{gridPoint}
\alias{gridPoint}
\title{#' R6 Class Representing a simulation grid point}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid point}

\item{\code{lon}}{Longitude of the grid point}

\item{\code{lat}}{Latitude of the grid point}

\item{\code{weather}}{Samara climate data generated from marksim}

\item{\code{simuweather}}{Samara climate data for simulation of given year}

\item{\code{result}}{Model simulation}

\item{\code{soilParam}}{soil parameters of grid point}

\item{\code{dateParam}}{sowing date of grid point}

\item{\code{parent}}{Grid parent of grid point}

\item{\code{test}}{Test attribute TODO: REMOVE}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gridPoint-new}{\code{gridPoint$new()}}
\item \href{#method-gridPoint-set_weather}{\code{gridPoint$set_weather()}}
\item \href{#method-gridPoint-set_soilParam}{\code{gridPoint$set_soilParam()}}
\item \href{#method-gridPoint-set_dateParam}{\code{gridPoint$set_dateParam()}}
\item \href{#method-gridPoint-genClimate}{\code{gridPoint$genClimate()}}
\item \href{#method-gridPoint-runSimulation}{\code{gridPoint$runSimulation()}}
\item \href{#method-gridPoint-clone}{\code{gridPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid point object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$new(parent = NA, name = "11", lon = -4.3, lat = 11.18)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Grid parent of grid point}

\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{lon}}{Longitude of grid point}

\item{\code{lat}}{Latitude of grid point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Change weather of grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_weather(weather)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weather}}{Weather dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_soilParam"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_soilParam}{}}}
\subsection{Method \code{set_soilParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_soilParam()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_dateParam"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_dateParam}{}}}
\subsection{Method \code{set_dateParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_dateParam(year)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Year of simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate at grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$genClimate(rcp, year, yearNb, modelNb, path, pathCLI)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-runSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-runSimulation}{}}}
\subsection{Method \code{runSimulation()}}{
Run simulation on grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$runSimulation(param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Parameter values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
