% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_models.R
\name{launch_sim}
\alias{launch_sim}
\title{Launch Crop Model simulation}
\usage{
launch_sim(model = "Samara", data, params)
}
\arguments{
\item{model}{One of the supported crop models (Samara or STICS)}

\item{data}{Either a path to the STICS workspace containing USM files
or a list of dataframes with input data (parameters, weather) for Samara}

\item{params}{Additional parameters to tune the model (see examples)}
}
\value{
A dataframe with the result of the simulation
}
\description{
Launch a simulation of one of the supported crop models using
input from dataframes/files for plant/soil parameters, and weather data
}
\examples{
param <- read.table("samara_parameters.csv")
weather <- read.table("samara_weather.csv")
resSamara <- launch_sim("Samara", param, weather)

simOptions <- stics_wrapper_option(javastics_path = "path_to_javastics",
workspace_path = "data/stics/", target_path = "examples_stics/results/",
verbose = false)
resStics <- launch_sim("STICS", "data/stics/", simOptions)
}
