% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error}
\usage{
MAE(Obs, Sim)
}
\arguments{
\item{Obs}{Vector (or single value) of observed value(s)}

\item{Sim}{Vector (or single value) of simulated value(s)}
}
\value{
The mean absolute error of \code{Obs} and \code{Sim}
}
\description{
Computes Mean Absolute Error between observed
and simulated values
}
\examples{
MAE(c(1,1,2), c(1,2,2))
MAE(2, 3)
}
