% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MSE}
\alias{MSE}
\title{Mean Squared Error}
\usage{
MSE(Obs, Sim)
}
\arguments{
\item{Obs}{Vector (or single value) of observed value(s)}

\item{Sim}{Vector (or single value) of simulated value(s)}
}
\value{
The mean squared error of \code{Obs} and \code{Sim}
}
\description{
Computes Mean Squared Error between observed and
simulated values
}
\examples{
MSE(c(1,1,2), c(1,2,2))
MSE(2, 3)
}
