% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rMAE}
\alias{rMAE}
\title{Relative Mean Absolute Error}
\usage{
rMAE(Obs, Sim)
}
\arguments{
\item{Obs}{Vector (or single value) of observed value(s)}

\item{Sim}{Vector (or single value) of simulated value(s)}
}
\value{
The relative mean absolute error of \code{Obs} and \code{Sim}
}
\description{
Computes Relative Mean Absolute Error between observed
and simulated values, i.e. the Mean Absolute Error divided by the mean of
observations (see \code{\link{MAE}})
}
\examples{
rMAE(c(1,1,2), c(1,2,2))
rMAE(2, 3)
}
