% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuf.R
\name{rmerge.all}
\alias{rmerge.all}
\title{Merge rows inside dataframe}
\usage{
rmerge.all(df, gb)
}
\arguments{
\item{df}{Dataframe with rows to merge}

\item{gb}{Group by column}
}
\value{
Dataframe with merged rows grouped by \code{gb}
}
\description{
Merge rows inside dataframe based on a group by.
(requires dplyr library)
(temporary fix, need to find better solution)
}
\examples{
df <- data.frame(Date = c(1,1,2,2,3),
                Mesure1 = c(NA,11,17,NA,13),
                Mesure2 = c(10,NA,NA,15,14))
dfMerged <- rmerge.all(df,"Date")
}
