% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuf.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Square Error}
\usage{
RMSE(Obs, Sim)
}
\arguments{
\item{Obs}{Vector (or single value) of observed value(s)}

\item{Sim}{Vector (or single value) of simulated value(s)}
}
\value{
The root mean square error of \code{Obs} and \code{Sim}
}
\description{
Computes Root Mean Square Error between observed and
simulated values, i.e. the squared root of Mean Squared Error
(see \code{\link{MSE}})
}
\examples{
RMSE(c(1,1,2), c(1,2,2))
RMSE(2, 3)
}
