% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuf.R
\name{rRMSE}
\alias{rRMSE}
\title{Relative Root Mean Square Error}
\usage{
rRMSE(Obs, Sim)
}
\arguments{
\item{Obs}{Vector (or single value) of observed value(s)}

\item{Sim}{Vector (or single value) of simulated value(s)}
}
\value{
The relative root mean square error of \code{Obs} and \code{Sim}
}
\description{
Computes Relative Root Mean Squared Error between observed and
simulated values, i.e. the Root Mean Square Error divided by the mean of
observations (see \code{\link{RMSE}})
}
\examples{
rRMSE(c(1,1,2), c(1,2,2))
rRMSE(2, 3)
}
