% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rsquared}
\alias{rsquared}
\title{Coefficient of determination (R squared)}
\usage{
rsquared(Obs, Sim)
}
\arguments{
\item{Obs}{Vector of observed value(s)}

\item{Sim}{Vector of simulated value(s)}
}
\value{
The R squared of \code{Obs} and \code{Sim}
}
\description{
Computes Coefficient of Determination between observed and
simulated values
}
\examples{
rsquared(c(1,1,2), c(1,2,2))
}
