% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimation.R
\name{sample_estimation}
\alias{sample_estimation}
\title{Sample parameter estimation for Samara}
\usage{
sample_estimation(maxiter = 200, seed = 2204, score_fn = get_score)
}
\arguments{
\item{maxiter}{Maximal number of iterations, default to 20 (see above)}

\item{seed}{Random seed value}
}
\description{
Simple estimation sample of one variety in one location using
a custom score function. Eight parameters are estimated considering
phenotypic data of 4 replicates. The estimation is performed using DEoptim,
see more details in \insertRef{deoptim}{CGMTPE}
For testing purpose the number of iteration is set to 20, but should be
higher in real situations. The size of the population is set by default
to be 10 times the number of estimated parameters.
See more info about convergence in \insertRef{hu2013}{CGMTPE}.
This sample script does not use parallelisation
}
\examples{
sample_estimation(20,1337, get_score)
}
