% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEa}
\alias{TPEa}
\title{#' R6 Class Representing a TPE analysis}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{model}}{Name of the crop model used for simulations}

\item{\code{varieties}}{List of varieties names used for TPE analysis}

\item{\code{environments}}{List of environment names used for TPE analysis}

\item{\code{genotypes}}{List of alternate varieties names used for TPE analysis}

\item{\code{grid}}{Simulation grid}

\item{\code{parameters}}{A vector or dataframe with all parameters used for
simulation with the corresponding `model`}

\item{\code{weathers}}{A dataframe or list of dataframe with weather data for
each of the `environments`}

\item{\code{observations}}{A dataframe or list of dataframe with observation *
data}

\item{\code{estimParam}}{Values of estimated parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEa-new}{\code{TPEa$new()}}
\item \href{#method-TPEa-set_model}{\code{TPEa$set_model()}}
\item \href{#method-TPEa-set_weather}{\code{TPEa$set_weather()}}
\item \href{#method-TPEa-set_obs}{\code{TPEa$set_obs()}}
\item \href{#method-TPEa-set_eparam}{\code{TPEa$set_eparam()}}
\item \href{#method-TPEa-update_param}{\code{TPEa$update_param()}}
\item \href{#method-TPEa-initMessage}{\code{TPEa$initMessage()}}
\item \href{#method-TPEa-createGrid}{\code{TPEa$createGrid()}}
\item \href{#method-TPEa-genClim}{\code{TPEa$genClim()}}
\item \href{#method-TPEa-runGridSim}{\code{TPEa$runGridSim()}}
\item \href{#method-TPEa-runEstimation}{\code{TPEa$runEstimation()}}
\item \href{#method-TPEa-clone}{\code{TPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$new(
  name = "TPEa_1",
  model = "Samara",
  varieties = NA,
  environments = NA,
  genotypes = NA,
  parameters = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{model}}{Name of the crop model used for simulations}

\item{\code{varieties}}{List of varieties}

\item{\code{environments}}{List of environments}

\item{\code{genotypes}}{Optional. List of alternate variety names}

\item{\code{parameters}}{A vector or dataframe with all crop model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set crop model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Dataframe (or list of df) of weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_obs"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_obs}{}}}
\subsection{Method \code{set_obs()}}{
Set observation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_obs(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Dataframe (or list of df) of observation data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_eparam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_eparam}{}}}
\subsection{Method \code{set_eparam()}}{
Set estimated parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_eparam(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Vector of parameter values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-update_param"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-update_param}{}}}
\subsection{Method \code{update_param()}}{
Update parameter with estimated parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$update_param(val, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Vector of parameter values}

\item{\code{names}}{Vector of parameter names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a simulation grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createGrid(
  name = "grid1",
  res = 5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{res}}{Resolution of the grid}

\item{\code{cols}}{Number of columns in the grid}

\item{\code{rows}}{Number of rows in the grid}

\item{\code{lon}}{Optional. Starting longitude of the grid}

\item{\code{lat}}{Optional. Starting latitude of the grid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-genClim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-genClim}{}}}
\subsection{Method \code{genClim()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$genClim(
  rcp = "rcp26",
  year = 2015,
  yearNb = 99,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runGridSim(row = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{row}}{Row of parameter dataframe to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runEstimation(
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxiter}}{Maximum number of iteration for DEoptim}

\item{\code{paramnames}}{Vector of parameter names to be estimated}

\item{\code{metric}}{Metric to use for fitness computation}

\item{\code{score_fn}}{Function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
