% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysis.R
\name{TPEase}
\alias{TPEase}
\title{R6 Class representing a TPE analysis and simulation environment}
\usage{
res <- TPEase$new("TPEase1", model, varieties, genotypes, vparameters,
                  environments, eparameters, eName, weathers, observations)
}
\description{
Main object containing all objects and functions used for parameter
estimation and TPE analysis
}
\details{
The \code{TPEase} object is the base building block of a
`TPE analysis and simulation environment` process, on which to built
additional bricks: crop model parameter estimation, statistical analysis
of simulation results, weather data generation, mapping and plotting, etc.
The \code{TPEase} can contain a list of \code{TPEaseVar} objects
(crop varieties) and a list of \code{TPEaseTPE} objects (TPE analyses)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The \code{TPEase} identifier}

\item{\code{model}}{The name of the crop model used in this \code{TPEase}
(only "Samara" is supported at the momemnt)}

\item{\code{varieties}}{A list of \code{TPEaseVar} objects}

\item{\code{TPEanalyses}}{A list of \code{TPEaseTPE} objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEase-new}{\code{TPEase$new()}}
\item \href{#method-TPEase-set_name}{\code{TPEase$set_name()}}
\item \href{#method-TPEase-set_model}{\code{TPEase$set_model()}}
\item \href{#method-TPEase-set_var}{\code{TPEase$set_var()}}
\item \href{#method-TPEase-get_varNames}{\code{TPEase$get_varNames()}}
\item \href{#method-TPEase-get_genotypes}{\code{TPEase$get_genotypes()}}
\item \href{#method-TPEase-get_varid}{\code{TPEase$get_varid()}}
\item \href{#method-TPEase-createTPE}{\code{TPEase$createTPE()}}
\item \href{#method-TPEase-createGrid}{\code{TPEase$createGrid()}}
\item \href{#method-TPEase-genClimate}{\code{TPEase$genClimate()}}
\item \href{#method-TPEase-runGridSim}{\code{TPEase$runGridSim()}}
\item \href{#method-TPEase-createMap}{\code{TPEase$createMap()}}
\item \href{#method-TPEase-runClustering}{\code{TPEase$runClustering()}}
\item \href{#method-TPEase-plotMap}{\code{TPEase$plotMap()}}
\item \href{#method-TPEase-addToPlot}{\code{TPEase$addToPlot()}}
\item \href{#method-TPEase-print_maps}{\code{TPEase$print_maps()}}
\item \href{#method-TPEase-runEstimation}{\code{TPEase$runEstimation()}}
\item \href{#method-TPEase-clone}{\code{TPEase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPEase object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$new(
  name = "TPEase_1",
  model = "Samara",
  varieties = NA,
  genotypes = NA,
  vparameters = NA,
  environments = NA,
  eparameters = NA,
  eName = "estim1",
  weathers = NA,
  observations = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of the TPEase}

\item{\code{model}}{A \code{character} string identifier of the crop model used}

\item{\code{varieties}}{A \code{vector} of varieties names}

\item{\code{genotypes}}{Optional. A \code{vector} of alternate varieties names}

\item{\code{vparameters}}{A \code{vector} or \code{data.frame}
with variety specific parameters}

\item{\code{environments}}{A \code{vector} of environments names}

\item{\code{eparameters}}{A \code{vector} or \code{data.frame}
with environment specific parameters}

\item{\code{eName}}{A \code{character} string identifier of the initial
\code{EstimVar} object}

\item{\code{weathers}}{A \code{list} (each entry is an environment)
of \code{data.frames} containing weather data}

\item{\code{observations}}{A \code{list} (each entry is a variety)
of \code{lists} (each entry is an environment) of \code{data.frames}
containing observations (of a given \code{TPEaseVar} [first key] in a
given \code{TPEaseEnv} [second key])}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEase` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPEase name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPEase name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set varieties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New \code{TPEaseVar} objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_varNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_varNames}{}}}
\subsection{Method \code{get_varNames()}}{
Get names of all TPEaseVar
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_varNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_genotypes}{}}}
\subsection{Method \code{get_genotypes()}}{
Get names of all genotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_genotypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_varid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_varid}{}}}
\subsection{Method \code{get_varid()}}{
Get TPEaseVar id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_varid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A variety identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createTPE"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createTPE}{}}}
\subsection{Method \code{createTPE()}}{
Create a \code{TPEaseTPE} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createTPE(name = "TPEa_1")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string of TPE name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a \code{TPEGrid} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createGrid(
  tpeID = 1,
  name = "g1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{TPEaseTPE} identifier on which to create grid}

\item{\code{name}}{A \code{character} string identifier of the grid}

\item{\code{varID}}{A \code{TPEaseVar} identifier to use for simulation
on the grid}

\item{\code{latres}}{A \code{numeric} value of the latitude resolution
of the grid}

\item{\code{lonres}}{A \code{numeric} value of the longitude resolution
of the grid}

\item{\code{cols}}{A \code{numeric} value of the number of columns in the grid}

\item{\code{rows}}{A \code{numeric} value of the number of rows in the grid}

\item{\code{lon}}{A \code{numeric} value of the starting longitude
of the grid in decimal degrees}

\item{\code{lat}}{A \code{numeric} value of the starting latitude
of the grid in decimal degrees}

\item{\code{multigrid}}{A \code{boolean }indicating if one grid should be
created for each genotype with the same \code{cols}, \code{rows},
\code{lon} and \code{lat}. If true, \code{varID} will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several \code{TPEGrid}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$genClimate(
  tpeID = 1,
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  seed = NA,
  filesE = F,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{TPEaseTPE} identifier of the grid's parent}

\item{\code{gridID}}{Optional. A \code{vector} of \code{TPEGrid} identifiers
(either index or name). By default will run all grids of the \code{tpeID}}

\item{\code{rcp}}{A \code{character} string of the name of the Representative
Concentration Pathway to use. One of the following options :
\code{"rcp26","rcp45","rcp60","rcp85"}}

\item{\code{year}}{A \code{numeric} value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A \code{numeric} value of the number of
replicates to simulate}

\item{\code{modelNb}}{A \code{character} string of the general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A \code{character} string of the path to the
marksim standalone (careful, this path can not contain any spaces)}

\item{\code{pathCLI}}{Optional. A \code{character} string of the path to
CLI folder. If no value is provided, the CLI folder will be considered
in the same folder as the marksim standalone}

\item{\code{seed}}{A \code{numeric} value of the seed to use to generate climate}

\item{\code{filesE}}{A \code{boolean} indicating ff weather files already exist
and should be used (existing weather files should be located in a
"weathers" subfolder of the \code{path})}

\item{\code{verbose}}{A \code{boolean} indicating ff messages about starting
climate generation should be shown}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several \code{TPEGrid}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runGridSim(
  tpeID = 1,
  gridID = NA,
  varID = NA,
  soilData = soil,
  latlonData = lat_lon,
  cumulP = cumul,
  traitList = NULL,
  savePath = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{TPEaseTPE} identifier on which to run
grid simulations}

\item{\code{gridID}}{Optional. A \code{vector} of \code{TPEGrid} identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A \code{TPEaseVar} identifier to use for simulation
on the grid. If NA, will use the variety attached to grid}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.
TODO: implement generic methods for soil, latlon, cumul}

\item{\code{traitList}}{Optional. A \code{vector} of trait names to extract from
simulations. This will delete the simulations and only keep the
maximum values for each year of the selected traits}

\item{\code{savePath}}{Optional. A \code{character} string of the path where to
save simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create \code{TPEMap} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createMap(tpeID = 1, name = "map1", bounds = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{TPEaseTPE} on which to create the map}

\item{\code{name}}{A \code{character} string identifier of the map}

\item{\code{bounds}}{Optional. A \code{vector} of four \code{numeric} values
as decimal degree of north, east, south, west bounds to crop
the world map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runClustering"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runClustering}{}}}
\subsection{Method \code{runClustering()}}{
Run TPE clustering by performing Principle Component
Analysis (PCA) and Hierarchical Clustering on Principle Component (HCPC)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runClustering(
  tpeID = 1,
  mapID = 1,
  traitList = c("GrainYieldPop"),
  nbDim = 5,
  nbClust = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{PEaseTPE} identifier parent of the \code{TPEMap}}

\item{\code{mapID}}{A value or list of values of \code{TPEMap} identifiers
(either index or names) on which to perform clustering}

\item{\code{traitList}}{A \code{vector} of trait names to be used for PCA}

\item{\code{nbDim}}{An \code{integer} of number of dimensions to use for PCA}

\item{\code{nbClust}}{An \code{integer} of number of clusters to use for HCPC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$plotMap(tpeID = 1, mapID = 1, trait = "cluster", isFactor = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{TPEeaseTPE} identifier parent of the \code{TPEMap}}

\item{\code{mapID}}{A \code{TPEMap} identifier to plot}

\item{\code{trait}}{A \code{character} string identifier of the data to plot,
by default will plot the cluster computed by the runClustering function
of the TPE analysis object}

\item{\code{isFactor}}{A \code{boolean} indicating if the trait should be
considered as a factor for plotting (cluster is a factor)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-addToPlot"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-addToPlot}{}}}
\subsection{Method \code{addToPlot()}}{
Add ggplot2 objects to existing plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$addToPlot(tpeID = 1, mapID = 1, plotID = 1, plotAdd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{numeric} value identifier of the \code{TPEaseTPE}}

\item{\code{mapID}}{A \code{TPEMap} identifier containing the plot}

\item{\code{plotID}}{A \code{numeric} value identifier of the plot}

\item{\code{plotAdd}}{A list of \code{ggplot2} objects to pass to the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-print_maps"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-print_maps}{}}}
\subsection{Method \code{print_maps()}}{
Print a given map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$print_maps(tpeID = 1, mapID = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A \code{numeric} value identifier of the \code{TPEaseTPE}}

\item{\code{mapID}}{A \code{numeric} value identifier of the \code{TPEMap}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation for a given variety
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runEstimation(
  varID = 1,
  estimID = 1,
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varID}}{A value or list of values of \code{TPEaseVar} identifier(s)
(either index or name)}

\item{\code{estimID}}{A \code{VarEstim} identifier (either index or name)}

\item{\code{maxiter}}{A \code{numeric} value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A \code{vector} of parameter names to be estimated}

\item{\code{metric}}{A \code{character} string with the name of the metric
to use for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A \code{function} to compute fitness,
see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment
TODO: implement weigh_fn}

\item{\code{bounds}}{A \code{matrix} or \code{data.frame} with lower (row1)
and upper (row2) bounds for each of the parameters in `paramnames` (cols)}

\item{\code{...}}{Additional parameters to be passed to \code{DEoptim.control}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
