% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_env.R
\name{TPEaseEnv}
\alias{TPEaseEnv}
\title{R6 Class representing a TPEaseEnv Environment}
\description{
The `TPEaseEnv` object contains all information of a given environment,
defined as a combination of : soil, weather and crop management data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEaseEnv}

\item{\code{parent}}{A TPEase object parent of the Environment}

\item{\code{weather}}{A EnvWeather object}

\item{\code{soil}}{A EnvSoil object}

\item{\code{cm}}{A EnvCM object}

\item{\code{parameters}}{A dataframe with all crop model environment parameters}

\item{\code{observations}}{A list of dataframes of observations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEaseEnv-new}{\code{TPEaseEnv$new()}}
\item \href{#method-TPEaseEnv-set_weather}{\code{TPEaseEnv$set_weather()}}
\item \href{#method-TPEaseEnv-set_param}{\code{TPEaseEnv$set_param()}}
\item \href{#method-TPEaseEnv-clone}{\code{TPEaseEnv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseEnv-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseEnv-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPEaseEnv object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseEnv$new(name = "e1", parent, parameters, weather, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEaseEnv}

\item{\code{parent}}{A TPEase object parent of the Environment}

\item{\code{parameters}}{A dataframe with all crop model environment parameters}

\item{\code{weather}}{A dataframe with weather data}

\item{\code{observations}}{A list of dataframes of observations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEenv` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseEnv-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseEnv-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseEnv$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe with weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseEnv-set_param"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseEnv-set_param}{}}}
\subsection{Method \code{set_param()}}{
Set parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseEnv$set_param(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe with parameter values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseEnv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseEnv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseEnv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
