% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEaseTPE}
\alias{TPEaseTPE}
\title{R6 Class representing a TPEaseTPE}
\description{
The `TPEaseTPE` object is an environment containing all objects used for
TPE analysis (grids and maps)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEaseTPE}

\item{\code{model}}{A character string identifier of the used crop model
(only "Samara" is supported for now)}

\item{\code{grids}}{A list of TPEGrid objects}

\item{\code{results}}{A dataframe with grid simulation result summary}

\item{\code{maps}}{A list of TPEMap objects for each of the TPEGrid}

\item{\code{bootstrapGP}}{A list of GridPoints to perform bootstrap}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEaseTPE-new}{\code{TPEaseTPE$new()}}
\item \href{#method-TPEaseTPE-initMessage}{\code{TPEaseTPE$initMessage()}}
\item \href{#method-TPEaseTPE-set_name}{\code{TPEaseTPE$set_name()}}
\item \href{#method-TPEaseTPE-set_model}{\code{TPEaseTPE$set_model()}}
\item \href{#method-TPEaseTPE-set_grid}{\code{TPEaseTPE$set_grid()}}
\item \href{#method-TPEaseTPE-set_map}{\code{TPEaseTPE$set_map()}}
\item \href{#method-TPEaseTPE-get_gridNames}{\code{TPEaseTPE$get_gridNames()}}
\item \href{#method-TPEaseTPE-get_mapNames}{\code{TPEaseTPE$get_mapNames()}}
\item \href{#method-TPEaseTPE-get_gridid}{\code{TPEaseTPE$get_gridid()}}
\item \href{#method-TPEaseTPE-get_mapid}{\code{TPEaseTPE$get_mapid()}}
\item \href{#method-TPEaseTPE-get_results}{\code{TPEaseTPE$get_results()}}
\item \href{#method-TPEaseTPE-createGrid}{\code{TPEaseTPE$createGrid()}}
\item \href{#method-TPEaseTPE-genClimate}{\code{TPEaseTPE$genClimate()}}
\item \href{#method-TPEaseTPE-runGridSim}{\code{TPEaseTPE$runGridSim()}}
\item \href{#method-TPEaseTPE-createMap}{\code{TPEaseTPE$createMap()}}
\item \href{#method-TPEaseTPE-runClustering}{\code{TPEaseTPE$runClustering()}}
\item \href{#method-TPEaseTPE-plotMap}{\code{TPEaseTPE$plotMap()}}
\item \href{#method-TPEaseTPE-addToPlot}{\code{TPEaseTPE$addToPlot()}}
\item \href{#method-TPEaseTPE-bootstrap}{\code{TPEaseTPE$bootstrap()}}
\item \href{#method-TPEaseTPE-clone}{\code{TPEaseTPE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPEaseTPE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$new(name = "TPE_1", model = "Samara", parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEaseTPE}

\item{\code{model}}{A character string identifier of the used crop model}

\item{\code{parent}}{A TPEase object parent of the TPEaseTPE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEaseTPE` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPEaseTPE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPEaseTPE name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPEaseTPE name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-set_grid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-set_grid}{}}}
\subsection{Method \code{set_grid()}}{
Set grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$set_grid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPEGrid objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-set_map"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-set_map}{}}}
\subsection{Method \code{set_map()}}{
Set maps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$set_map(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPEMap objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-get_gridNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-get_gridNames}{}}}
\subsection{Method \code{get_gridNames()}}{
Get names of all TPEGrid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$get_gridNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-get_mapNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-get_mapNames}{}}}
\subsection{Method \code{get_mapNames()}}{
Get names of all TPEMap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$get_mapNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-get_gridid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-get_gridid}{}}}
\subsection{Method \code{get_gridid()}}{
Get TPEGrid id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$get_gridid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A grid identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-get_mapid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-get_mapid}{}}}
\subsection{Method \code{get_mapid()}}{
Get TPEMap id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$get_mapid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A map identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get simulation results of all grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$get_results(traitList = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{A list of variables to extract as results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a TPEGrid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$createGrid(
  name = "g1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the grid}

\item{\code{varID}}{A TPEaseVar identifier to use for simulation on the grid}

\item{\code{latres}}{A numeric value of the latitude resolution of the grid}

\item{\code{lonres}}{A numeric value of the longitude resolution of the grid}

\item{\code{cols}}{A numeric value of the number of columns in the grid}

\item{\code{rows}}{A numeric value of the number of rows in the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the grid in
decimal degrees}

\item{\code{lat}}{A numeric value of the starting latitude of the grid in
decimal degrees}

\item{\code{multigrid}}{A boolean indicating if one grid should be created
for each TPEaseVar with the same cols, rows, lon and lat. If true, varID
will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several TPEGrid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$genClimate(
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  filesE = F,
  verbose = T,
  seed = 1337
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the  general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{filesE}}{A boolean indicating if weather files already exist}

\item{\code{verbose}}{A boolean indicating ff messages about starting
climate generation should be shown}

\item{\code{seed}}{An integer number to use as seed for marksim weather
generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several TPEGrid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$runGridSim(
  gridID = NA,
  varID = NA,
  soilData = soil,
  latlonData = lat_lon,
  cumulP = cumul,
  traitList = NULL,
  savePath = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A TPeaseVar identifier to use for
simulation on the grid. If NA, will use the variety attached to grid}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. A vector of trait names to extract from
simulations. This will delete the simulations and only keep the max for
each year}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create TPEMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$createMap(name = "map1", bounds = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the map}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map}

\item{\code{res}}{Not used at the moment.
TODO: remove map res}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-runClustering"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-runClustering}{}}}
\subsection{Method \code{runClustering()}}{
Run TPE clustering by performing Principle Component
Analysis (PCA) and Hierarchical Clustering on Principle Component (HCPC)
on a TPEMap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$runClustering(
  mapID = 1,
  traitList = c("GrainYieldPop"),
  nbDim = 5,
  nbClust = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A value or list of values of map identifiers
(either index or names) on which to perform clustering}

\item{\code{traitList}}{A vector of variable names to be used for PCA}

\item{\code{nbDim}}{An integer of number of dimensions to use for PCA}

\item{\code{nbClust}}{An integer of number of clusters to use for HCPC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create a plot on a TPEMap based on TPEGrid simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$plotMap(mapID = 1, trait, isFactor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A value of map identifier}

\item{\code{trait}}{A character string identifier of the data to plot,
by default will plot the cluster computed by the runClustering function
of the TPEeaseTPE object}

\item{\code{isFactor}}{A boolean indicating if the trait should be considered
as a factor for plotting (cluster is a factor)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-addToPlot"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-addToPlot}{}}}
\subsection{Method \code{addToPlot()}}{
Add ggplot2 objects to existing plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$addToPlot(mapID = 1, plotID = 1, plotAdd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A numeric value identifier of the TPEMap}

\item{\code{plotID}}{A numeric value identifier of the plot}

\item{\code{plotAdd}}{A list of ggplot2 objects to pass to the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-bootstrap}{}}}
\subsection{Method \code{bootstrap()}}{
Perform Bootstrap Analysis to test if gridded simulations
are close to simulations on observed sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$bootstrap(
  nameList,
  latList,
  lonList,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  seed = 1337,
  soilData,
  latlonData,
  cumulP,
  traitList,
  savePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nameList}}{A list of names of sites to test}

\item{\code{latList}}{A list of latitudes of sites to test}

\item{\code{lonList}}{A list of longitudes of sites to test}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of replicates to simulate}

\item{\code{modelNb}}{A character string of the general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{seed}}{An integer number to use as seed for marksim weather
generator}

\item{\code{soilData}}{Tmp fix for Adam et al.}

\item{\code{latlonData}}{Tmp fix for Adam et al.}

\item{\code{cumulP}}{Tmp fix for Adam et al.}

\item{\code{traitList}}{A vector of variable names}

\item{\code{savePath}}{A character string of the path to save bootstrap
simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseTPE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseTPE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseTPE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
