% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_cm.R
\name{TPEcm}
\alias{TPEcm}
\title{#' R6 Class Representing a crop managment}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the crop management}

\item{\code{parameters}}{A dataframe with crop management parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEcm-new}{\code{TPEcm$new()}}
\item \href{#method-TPEcm-set_cmParam}{\code{TPEcm$set_cmParam()}}
\item \href{#method-TPEcm-set_dateparam}{\code{TPEcm$set_dateparam()}}
\item \href{#method-TPEcm-set_latlon}{\code{TPEcm$set_latlon()}}
\item \href{#method-TPEcm-clone}{\code{TPEcm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-new}{}}}
\subsection{Method \code{new()}}{
Create a new variety object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$new(name = "e1", parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEenv` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-set_cmParam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-set_cmParam}{}}}
\subsection{Method \code{set_cmParam()}}{
Set crop management parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$set_cmParam(name, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} value (or \code{vector}) of
crop management parameter name(s)}

\item{\code{val}}{A \code{numeric} (or \code{vector}) of crop management
parameter values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-set_dateparam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-set_dateparam}{}}}
\subsection{Method \code{set_dateparam()}}{
Set date parameters (starting, ending and sowing date)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$set_dateparam(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A vector of size 3 with starting, ending and sowing dates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-set_latlon"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-set_latlon}{}}}
\subsection{Method \code{set_latlon()}}{
Set date latitude and longitude
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$set_latlon(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A vector of size 2 with latitude and longitude of the site}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
