% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_env.R
\name{TPEenv}
\alias{TPEenv}
\title{#' R6 Class Representing a TPE analysis}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{parent}}{TPE analysis parent}

\item{\code{weather}}{A dataframe with all weather data used for simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEenv-new}{\code{TPEenv$new()}}
\item \href{#method-TPEenv-set_weather}{\code{TPEenv$set_weather()}}
\item \href{#method-TPEenv-clone}{\code{TPEenv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEenv-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEenv-new}{}}}
\subsection{Method \code{new()}}{
Create a new variety object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEenv$new(name = "e1", parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEenv` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEenv-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEenv-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEenv$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe with weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEenv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEenv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEenv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
