% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_map.R
\name{TPEmap}
\alias{TPEmap}
\title{#' R6 Class Representing a TPE map}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the map}

\item{\code{data}}{A SpatialPolygondataframe of the desired map}

\item{\code{PCAres}}{Principle Component Analysis result for `self$grid$gridRes`}

\item{\code{HCPCres}}{Hierarchical Clustering on Principle Components result for
`PCAres`}

\item{\code{plots}}{A list of ggplot2 plots from the map}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEmap-new}{\code{TPEmap$new()}}
\item \href{#method-TPEmap-runPCA}{\code{TPEmap$runPCA()}}
\item \href{#method-TPEmap-runHCPC}{\code{TPEmap$runHCPC()}}
\item \href{#method-TPEmap-plotMap}{\code{TPEmap$plotMap()}}
\item \href{#method-TPEmap-clone}{\code{TPEmap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE map object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$new(name = "map1", bounds = NA, parent = NA, res = 150)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE map}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map}

\item{\code{parent}}{TPE parent}

\item{\code{res}}{Not active for now. A numeric value in sec of the resolution
of world map to use. Options are c(1, 150, 900) for respectively
30sec, 2.5min, 15min}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEmap` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-runPCA"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-runPCA}{}}}
\subsection{Method \code{runPCA()}}{
Run Principle Component Analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$runPCA(dfPCA = NULL, nbDim = 5, traitList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dfPCA}}{Data to perform PCA}

\item{\code{nbDim}}{Number of dimensions}

\item{\code{traitList}}{Vector of trait names to use for PCA}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-runHCPC"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-runHCPC}{}}}
\subsection{Method \code{runHCPC()}}{
Run Hierarchical Clustering on Principle Components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$runHCPC(nbClust)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nbClust}}{Number of clusters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$plotMap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
