% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soil.R
\name{TPEsoil}
\alias{TPEsoil}
\title{#' R6 Class Representing a TPE soil}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the weather}

\item{\code{parent}}{Parent environment}

\item{\code{parameters}}{Soil parameters}

\item{\code{soilParam}}{Boolean indicating if soil parameters were extracted
from HC27}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEsoil-new}{\code{TPEsoil$new()}}
\item \href{#method-TPEsoil-set_soilParam}{\code{TPEsoil$set_soilParam()}}
\item \href{#method-TPEsoil-clone}{\code{TPEsoil$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEsoil-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEsoil-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE soil object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEsoil$new(name = "soil1", parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE soil}

\item{\code{parent}}{Parent environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEsoil` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEsoil-set_soilParam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEsoil-set_soilParam}{}}}
\subsection{Method \code{set_soilParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEsoil$set_soilParam(soil, lat_lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soil}}{Dataframe with soil characteristics}

\item{\code{lat_lon}}{Dataframe with correspondance of lat/lon with soil df}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEsoil-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEsoil-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEsoil$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
