% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_weather.R
\name{TPEweather}
\alias{TPEweather}
\title{#' R6 Class Representing a TPE weather}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the weather}

\item{\code{wData}}{A dataframe with weather data}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEweather-new}{\code{TPEweather$new()}}
\item \href{#method-TPEweather-clone}{\code{TPEweather$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE weather object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$new(name = "map1")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE weather}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEweather` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
