% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_estim.R
\name{VarEstim}
\alias{VarEstim}
\title{R6 Class representing a VarEstim Estimation process}
\usage{
res <- TPEase$new(...)
res$varieties[[1]]$createEstim(name, self, environments, eparam, weathers,
                               observations)
}
\description{
Estimation object for a given variety containing all information about
crop model parameter estimation
}
\details{
The \code{VarEstim} object contains all information to perform parameter
estimation for a given \code{TPEaseVar} variety including calibration
environments
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The \code{VarEstim} identifier}

\item{\code{parent}}{The \code{TPEaseVar} object parent of VarEstim}

\item{\code{environments}}{A \code{list} of \code{EstimEnv} objects to use for
estimation}

\item{\code{parameters}}{All parameters (variety x environment) of the crop
model}

\item{\code{DEparams}}{A \code{DEoptim.control} object with hyperparameters
used for \code{DEoptim} parameter estimation}

\item{\code{DEresult}}{The \code{DEoptim} result}

\item{\code{paramNames}}{The names of parameters to be estimated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VarEstim-new}{\code{VarEstim$new()}}
\item \href{#method-VarEstim-createEnv}{\code{VarEstim$createEnv()}}
\item \href{#method-VarEstim-runEstimation}{\code{VarEstim$runEstimation()}}
\item \href{#method-VarEstim-clone}{\code{VarEstim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-new"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{VarEstim} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$new(
  name = "estim1",
  environments,
  eparam,
  weathers,
  observations,
  parent
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of \code{VarEstim}}

\item{\code{environments}}{A \code{vector} of environments names}

\item{\code{eparam}}{A \code{data.frame} of \code{environments} parameters}

\item{\code{weathers}}{A \code{data.frame} of \code{environments} weathers}

\item{\code{observations}}{A \code{data.frame} of \code{environments}
observations for the parent \code{TPEaseVar}}

\item{\code{parent}}{A \code{TPEaseVar} parent object of \code{VarEstim}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VarEstim} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-createEnv"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-createEnv}{}}}
\subsection{Method \code{createEnv()}}{
Create a \code{EstimEnv} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$createEnv(name, eparam, weather, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of the \code{EstimEnv}}

\item{\code{eparam}}{A \code{data.frame} of environment parameters}

\item{\code{weather}}{A \code{data.frame} of environment's weather data}

\item{\code{observations}}{A \code{data.frame} of observations of the parent
\code{TPEaseVar} in the \code{EstimEnv}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$runEstimation(
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  bounds = NA,
  args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for \code{DEoptim}}

\item{\code{paramnames}}{A \code{vector} of parameter names to be estimated}

\item{\code{metric}}{A \code{character} string with the name of the metric to
use for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A \code{function} to compute fitness,
see \code{get_score}}

\item{\code{bounds}}{A \code{matrix} or \code{data.frame} with lower (col1)
and upper (col2) bounds for each of the parameters in \code{paramnames}
(rows)}

\item{\code{args}}{Additional parameters to be passed to \code{DEoptim.control}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
