% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_models.R
\name{construct_data}
\alias{construct_data}
\title{Construct Samara dataframes}
\usage{
construct_data(variety, genotype, itk)
}
\arguments{
\item{variety}{Name of the studied variety}

\item{genotype}{Corresponding genotype code of the variety}

\item{itk}{Name of the studied itk}

\item{workspace}{Path to the parent folder containing the subfolders obs,
params, meteos}
}
\value{
A list with parameter, weather and observation data
}
\description{
Construct dataframes of parameters and weather data from files
of individual variety and itk parameters to run a Samara simulation
}
\examples{
dfList <- construct_data("621B","G7","BAMA2014S1","examples_samara/")
param <- dfList[["parameters"]]
weather <- dfList[["weathers"]]
obs <- dfList[["observations"]]
}
