% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_tpe.R
\name{create_tpe}
\alias{create_tpe}
\title{Create a TPE analysis object}
\usage{
create_tpe(
  name = "TPEa_1",
  model = "Samara",
  varieties = NA,
  environments = NA,
  latStart = NA,
  lonStart = NA,
  genotypes = NA
)
}
\arguments{
\item{name}{test}

\item{model}{test}

\item{varieties}{test}

\item{environments}{test}

\item{latStart}{test}

\item{lonStart}{test}

\item{genotypes}{test}
}
\value{
A R6 class TPE analysis object
}
\description{
Create a TPE analysis object (from R6 class TPEa)
}
\examples{
TPE_analysis <- create_tpe("tpea1","Samara")
}
