% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_estim.R
\name{estimP}
\alias{estimP}
\title{#' R6 Class Representing a parameter estimation process}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of estimation}

\item{\code{parent}}{Parent variety}

\item{\code{environments}}{List of environments to use for estimation}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-estimP-new}{\code{estimP$new()}}
\item \href{#method-estimP-createEnv}{\code{estimP$createEnv()}}
\item \href{#method-estimP-clone}{\code{estimP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-new"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE soil object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$new(name = "estim1", parent, environments, eparam, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of estimation}

\item{\code{parent}}{Parent variety}

\item{\code{environments}}{List of environment names}

\item{\code{eparam}}{Parameters of environment}

\item{\code{observations}}{Observations of environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEsoil` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-createEnv"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-createEnv}{}}}
\subsection{Method \code{createEnv()}}{
Create environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$createEnv(name, eparam, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of environment}

\item{\code{eparam}}{Parameters of environment}

\item{\code{observations}}{Observations of environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-estimP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{estimP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
