% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_estimation.R
\name{get_score}
\alias{get_score}
\title{Compute fitness score between observations and simulations}
\usage{
get_score(obs, sim, metric = NA, weight_fn = NA)
}
\arguments{
\item{obs}{Dataframe of observations}

\item{sim}{Dataframe of simulations}

\item{metric}{Either a single value of the chosen metric to compute fitness
(RMSE, MAE) or a vector with a metric for each column of \code{obs}. TODO}

\item{weight_fn}{Optionnal. TODO}
}
\description{
TODO
}
\examples{
get_score(obs, sim, metric, weight_fn)
}
