% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gridpoint.R
\name{GridPoint}
\alias{GridPoint}
\title{#' R6 Class representing a simulation grid point}
\usage{
\code{GridPoint} are automatically created when creating a \code{TPEGrid}
}
\description{
Environment object of a \code{TPEGrid} containing all information about a
given environment used in the TPE analysis process
}
\details{
The \code{GridPoint} object is a subclass of \code{TPEaseEnv} specific for a
TPE analysis and containing all information of the Environment,
defined as a combination of : soil, weather and crop management data as well
as geographical location inside the \code{TPEGrid}
}
\section{Super class}{
\code{\link[CGMTPE:TPEaseEnv]{CGMTPE::TPEaseEnv}} -> \code{GridPoint}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lon}}{The longitude of the grid point}

\item{\code{lat}}{The latitude of the grid point}

\item{\code{result}}{Mean simulation results}

\item{\code{simulations}}{Raw simulation results}

\item{\code{gridParent}}{The \code{TPEGrid} parent object of \code{GridPoint}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GridPoint-new}{\code{GridPoint$new()}}
\item \href{#method-GridPoint-set_soilParam}{\code{GridPoint$set_soilParam()}}
\item \href{#method-GridPoint-set_dateParam}{\code{GridPoint$set_dateParam()}}
\item \href{#method-GridPoint-set_variety}{\code{GridPoint$set_variety()}}
\item \href{#method-GridPoint-set_weather}{\code{GridPoint$set_weather()}}
\item \href{#method-GridPoint-get_sim}{\code{GridPoint$get_sim()}}
\item \href{#method-GridPoint-get_meansd}{\code{GridPoint$get_meansd()}}
\item \href{#method-GridPoint-genClimate}{\code{GridPoint$genClimate()}}
\item \href{#method-GridPoint-runSimulation}{\code{GridPoint$runSimulation()}}
\item \href{#method-GridPoint-runBSSimulation}{\code{GridPoint$runBSSimulation()}}
\item \href{#method-GridPoint-clone}{\code{GridPoint$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="CGMTPE" data-topic="TPEaseEnv" data-id="set_param"><a href='../../CGMTPE/html/TPEaseEnv.html#method-TPEaseEnv-set_param'><code>CGMTPE::TPEaseEnv$set_param()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{GridPoint} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$new(parent = NA, name = "11", lon = -4.3, lat = 11.18)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{A \code{TPEGrid} parent object of the \code{GridPoint}}

\item{\code{name}}{A \code{character} string identifier of the \code{GridPoint}}

\item{\code{lon}}{A \code{numeric} value of the longitude of grid point
(in decimal degrees)}

\item{\code{lat}}{A \code{numeric} value of the latitude of grid point
(in decimal degrees)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GridPoint} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-set_soilParam"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-set_soilParam}{}}}
\subsection{Method \code{set_soilParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$set_soilParam(soil, lat_lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soil}}{Dataframe with soil characteristics}

\item{\code{lat_lon}}{Dataframe with correspondance of lat/lon with soil df}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-set_dateParam"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-set_dateParam}{}}}
\subsection{Method \code{set_dateParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$set_dateParam(year, cumulP, run = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Year of simulation}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{run}}{Tmp for Adam et al.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-set_variety"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-set_variety}{}}}
\subsection{Method \code{set_variety()}}{
Set variety
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$set_variety(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{\code{TPEaseVar} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{data.frame} of weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-get_sim"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-get_sim}{}}}
\subsection{Method \code{get_sim()}}{
Get simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$get_sim(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{character} string of the variable name to retrieve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-get_meansd"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-get_meansd}{}}}
\subsection{Method \code{get_meansd()}}{
Get mean and standard deviation of simulation result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$get_meansd(traitList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{A \code{vector} of trait names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$genClimate(rcp, year, yearNb, modelNb, path, pathCLI, seed, bs = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{A \code{character} string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A \code{numeric} value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A \code{numeric} value of the number of years to simulate}

\item{\code{modelNb}}{A \code{character} string of the  general circulation
model identifier to use, see \code{generateClimate}}

\item{\code{path}}{A \code{character} string with the path to the marksim
standalone. This path can not contain spaces}

\item{\code{pathCLI}}{Optional. A \code{character} string with the path to
CLI folder. If no value is provided, the CLI folder will be considered
in the same folder as the marksim standalone. This path can not
contain spaces}

\item{\code{seed}}{An \code{integer} number to use as seed for marksim weather
generator}

\item{\code{bs}}{A \code{boolean} indicating if should weathers be stored for
bootstrap}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-runSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-runSimulation}{}}}
\subsection{Method \code{runSimulation()}}{
Run simulation on \code{GridPoint}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$runSimulation(
  param,
  i,
  j,
  soilData,
  latlonData,
  cumulP,
  traitList,
  savePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A \code{data.frame} of all crop model parameter values}

\item{\code{i}}{An \code{integer} value of the row position of \code{GridPoint}
in the \code{TPEGrid}}

\item{\code{j}}{An \code{integer} value of the col position of \code{GridPoint}
in the \code{TPEGrid}}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. A \code{vector} of trait names to extract
from simulations. This will delete the simulations and only keep the max
for each year}

\item{\code{savePath}}{Optional. A \code{character} string of the path where to
save simulation files. If \code{NULL} (default),
will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-runBSSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-runBSSimulation}{}}}
\subsection{Method \code{runBSSimulation()}}{
Run simulation on \code{GridPoint} for bootstrap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$runBSSimulation(
  param,
  soilData,
  latlonData,
  cumulP,
  traitList,
  startingYear,
  savePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A \code{data.frame} of all crop model parameter values}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. A \code{vector} of trait names to extract
from simulations. This will delete the simulations and only keep the max
for each year}

\item{\code{startingYear}}{An \code{integer} value of the starting year of
bootstrap climate generation}

\item{\code{savePath}}{Optional. A \code{character} string of the path where to
save simulation files. If \code{NULL} (default),
will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GridPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GridPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GridPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
