% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gridpoint.R
\name{gridPoint}
\alias{gridPoint}
\title{#' R6 Class Representing a simulation grid point}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid point}

\item{\code{lon}}{Longitude of the grid point}

\item{\code{lat}}{Latitude of the grid point}

\item{\code{parameters}}{Combination of variety and environment parameters}

\item{\code{weather}}{A weather object}

\item{\code{soil}}{A soil object}

\item{\code{cm}}{A soil object}

\item{\code{result}}{Mean results}

\item{\code{simulations}}{Model simulations}

\item{\code{parent}}{Grid parent of grid point}

\item{\code{variety}}{Optional. Specific variety for given gridPoint}

\item{\code{test}}{Test attribute}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gridPoint-new}{\code{gridPoint$new()}}
\item \href{#method-gridPoint-set_soilParam}{\code{gridPoint$set_soilParam()}}
\item \href{#method-gridPoint-set_dateParam}{\code{gridPoint$set_dateParam()}}
\item \href{#method-gridPoint-set_variety}{\code{gridPoint$set_variety()}}
\item \href{#method-gridPoint-set_weather}{\code{gridPoint$set_weather()}}
\item \href{#method-gridPoint-get_sim}{\code{gridPoint$get_sim()}}
\item \href{#method-gridPoint-get_meansd}{\code{gridPoint$get_meansd()}}
\item \href{#method-gridPoint-genClimate}{\code{gridPoint$genClimate()}}
\item \href{#method-gridPoint-runSimulation}{\code{gridPoint$runSimulation()}}
\item \href{#method-gridPoint-runBSSimulation}{\code{gridPoint$runBSSimulation()}}
\item \href{#method-gridPoint-clone}{\code{gridPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid point object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$new(parent = NA, name = "11", lon = -4.3, lat = 11.18)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Grid parent of grid point}

\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{lon}}{Longitude of grid point}

\item{\code{lat}}{Latitude of grid point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_soilParam"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_soilParam}{}}}
\subsection{Method \code{set_soilParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_soilParam(soil, lat_lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soil}}{Dataframe with soil characteristics}

\item{\code{lat_lon}}{Dataframe with correspondance of lat/lon with soil df}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_dateParam"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_dateParam}{}}}
\subsection{Method \code{set_dateParam()}}{
Set soil parameters for grid point location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_dateParam(year, cumulP, run = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Year of simulation}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{run}}{Tmp for Adam et al.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_variety"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_variety}{}}}
\subsection{Method \code{set_variety()}}{
Set variety to this gridPoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_variety(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Object of type variety}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe of weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-get_sim"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-get_sim}{}}}
\subsection{Method \code{get_sim()}}{
Get simulation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$get_sim(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Variable name to retrieve}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-get_meansd"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-get_meansd}{}}}
\subsection{Method \code{get_meansd()}}{
Get mean and sd of simulation result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$get_meansd(traitList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{Vector of trait names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$genClimate(rcp, year, yearNb, modelNb, path, pathCLI, seed, bs = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}

\item{\code{seed}}{Integer number to use as seed for marksim weather generator}

\item{\code{bs}}{Boolean. Should weathers be stored for bootstrap}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-runSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-runSimulation}{}}}
\subsection{Method \code{runSimulation()}}{
Run simulation on grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$runSimulation(
  param,
  i,
  j,
  soilData,
  latlonData,
  cumulP,
  traitList,
  savePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Parameter values}

\item{\code{i}}{row position of grid point}

\item{\code{j}}{col position of grid point}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the max for
each year}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-runBSSimulation"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-runBSSimulation}{}}}
\subsection{Method \code{runBSSimulation()}}{
Run simulation on grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$runBSSimulation(
  param,
  soilData,
  latlonData,
  cumulP,
  traitList,
  startingYear,
  savePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Parameter values}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the max for
each year}

\item{\code{startingYear}}{Starting year of bootstrap climate gen}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gridPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gridPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gridPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
