% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEa}
\alias{TPEa}
\title{#' R6 Class Representing a TPE analysis}
\description{
The TPE Analysis object "TPEa" is an environment containing all objects
used for a TPE analysis (grid, maps, etc.)
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}

\item{\code{model}}{A character string with the name of the crop model
used for simulations (only "Samara" is supported for now)}

\item{\code{grids}}{A vector of simulation grids}

\item{\code{results}}{A dataframe with grid simulation results}

\item{\code{maps}}{A list of raster maps for each of the `grids`}

\item{\code{bootstrapGP}}{A list of gridpoints to perform bootstrap}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEa-new}{\code{TPEa$new()}}
\item \href{#method-TPEa-initMessage}{\code{TPEa$initMessage()}}
\item \href{#method-TPEa-set_name}{\code{TPEa$set_name()}}
\item \href{#method-TPEa-set_model}{\code{TPEa$set_model()}}
\item \href{#method-TPEa-set_grid}{\code{TPEa$set_grid()}}
\item \href{#method-TPEa-set_map}{\code{TPEa$set_map()}}
\item \href{#method-TPEa-get_gridNames}{\code{TPEa$get_gridNames()}}
\item \href{#method-TPEa-get_mapNames}{\code{TPEa$get_mapNames()}}
\item \href{#method-TPEa-get_gridid}{\code{TPEa$get_gridid()}}
\item \href{#method-TPEa-get_mapid}{\code{TPEa$get_mapid()}}
\item \href{#method-TPEa-get_results}{\code{TPEa$get_results()}}
\item \href{#method-TPEa-createGrid}{\code{TPEa$createGrid()}}
\item \href{#method-TPEa-genClimate}{\code{TPEa$genClimate()}}
\item \href{#method-TPEa-runGridSim}{\code{TPEa$runGridSim()}}
\item \href{#method-TPEa-createMap}{\code{TPEa$createMap()}}
\item \href{#method-TPEa-runClustering}{\code{TPEa$runClustering()}}
\item \href{#method-TPEa-plotMap}{\code{TPEa$plotMap()}}
\item \href{#method-TPEa-bootstrap}{\code{TPEa$bootstrap()}}
\item \href{#method-TPEa-clone}{\code{TPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$new(name = "TPEa_1", model = "Samara", parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}

\item{\code{model}}{A character string with the name of the crop model used
for simulations}

\item{\code{parent}}{Parent CGMTPE analysis object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPE analysis name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPE analysis name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_grid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_grid}{}}}
\subsection{Method \code{set_grid()}}{
Set grids, this is a list of grid objects,
use with caution this will overwrite existing grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_grid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New grids}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_map"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_map}{}}}
\subsection{Method \code{set_map()}}{
Set maps, this is a list of map objects,
use with caution, this will overwrite existing maps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_map(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New maps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_gridNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_gridNames}{}}}
\subsection{Method \code{get_gridNames()}}{
Get names of all grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_gridNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_mapNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_mapNames}{}}}
\subsection{Method \code{get_mapNames()}}{
Get names of all maps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_mapNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_gridid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_gridid}{}}}
\subsection{Method \code{get_gridid()}}{
Get grid id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_gridid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either grid name or a grid id (will return the grid id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_mapid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_mapid}{}}}
\subsection{Method \code{get_mapid()}}{
Get map id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_mapid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either map name or a map id (will return the map id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get simulation results of all grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_results(traitList = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{List of variables to extract as results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a simulation grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createGrid(
  name = "g1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the grid}

\item{\code{varID}}{A variety identifier to use for simulation on the grid}

\item{\code{latres}}{A numeric value of the latitude resolution of the grid}

\item{\code{lonres}}{A numeric value of the longitude resolution of the grid}

\item{\code{cols}}{A numeric value of the number of columns in the grid}

\item{\code{rows}}{A numeric value of the number of rows in the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the grid in
decimal degrees}

\item{\code{lat}}{A numeric value of the starting latitude of the grid in
decimal degrees}

\item{\code{multigrid}}{A boolean indicating if one grid should be created
for each genotype with the same cols, rows, lon and lat. If true, varID
will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$genClimate(
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  filesE = F,
  verbose = T,
  seed = 1337
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the  general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{filesE}}{Boolean. If weather files already exist}

\item{\code{verbose}}{Boolean. If messages about starting climate generation
should be shown}

\item{\code{seed}}{Integer number to use as seed for marksim weather generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runGridSim(
  gridID = NA,
  varID = NA,
  soilData = soil,
  latlonData = lat_lon,
  cumulP = cumul,
  traitList = NULL,
  savePath = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A variety identifier to use for
simulation on the grid. If NA, will use the variety attached to grid}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the max for
each year}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create raster map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createMap(name = "map1", bounds = NA, res = 150)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string defining the name of the map}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map}

\item{\code{res}}{Not used at the moment. A numeric value in sec of the
resolution of world map to use. Options are c(1, 150, 900)
for respectively 30sec, 2.5min, 15min}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runClustering"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runClustering}{}}}
\subsection{Method \code{runClustering()}}{
Run TPE clustering by performing Principle Component
Analysis (PCA) and Hierarchical Clustering on Principle Component (HCPC)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runClustering(
  mapID = 1,
  traitList = c("GrainYieldPop"),
  nbDim = 5,
  nbClust = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A value or list of values of map identifiers
(either index or names)on which to perform clustering}

\item{\code{traitList}}{Vector of variable names to be used for PCA}

\item{\code{nbDim}}{Integer of number of dimensions to use for PCA}

\item{\code{nbClust}}{Integer of number of clusters to use for HCPC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$plotMap(mapID = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{Id of map to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-bootstrap}{}}}
\subsection{Method \code{bootstrap()}}{
Bootstrap to test if gridded simulations are close to
simulations on observed sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$bootstrap(
  nameList,
  latList,
  lonList,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  seed = 1337,
  soilData,
  latlonData,
  cumulP,
  traitList
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nameList}}{List of names of sites to test}

\item{\code{latList}}{List of latitudes of sites to test}

\item{\code{lonList}}{List of longitudes of sites to test}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the  general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{seed}}{Integer number to use as seed for marksim weather generator}

\item{\code{soilData}}{Tmp fix for Adam et al.}

\item{\code{latlonData}}{Tmp fix for Adam et al.}

\item{\code{cumulP}}{Tmp fix for Adam et al.}

\item{\code{traitList}}{Vector of variable names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
