% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEa}
\alias{TPEa}
\title{#' R6 Class Representing a TPE analysis}
\description{
TODO
}
\details{
TODO
}
\examples{

## ------------------------------------------------
## Method `TPEa$set_model`
## ------------------------------------------------

TPE_analysis <- TPEa("TPEa1", "STICS")
TPE_analysis$model
TPE_analysis$set_model("Samara")
TPE_analysis$model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{model}}{Name of the crop model used for simulations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TPEa$new()}}
\item \href{#method-set_model}{\code{TPEa$set_model()}}
\item \href{#method-clone}{\code{TPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$new(name = NA, model = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name Identifier of the TPE analysis}

\item{\code{model}}{model Name of the crop model used for simulations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-set_model}{}}}
\subsection{Method \code{set_model()}}{
Change crop model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{TPE_analysis <- TPEa("TPEa1", "STICS")
TPE_analysis$model
TPE_analysis$set_model("Samara")
TPE_analysis$model
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
