% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEa}
\alias{TPEa}
\title{#' R6 Class Representing a TPE analysis}
\description{
The TPE Analysis object "TPEa" is an environment containing all objects
used for a TPE analysis (varieties, environments, maps, etc.)
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}

\item{\code{model}}{A character string with the name of the crop model
used for simulations (only "Samara" is supported for now)}

\item{\code{varieties}}{A list of varieties objects}

\item{\code{environments}}{A vector of environment names used for calibration}

\item{\code{grids}}{A vector of simulation grids}

\item{\code{maps}}{A list of raster maps for each of the `grids`}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEa-new}{\code{TPEa$new()}}
\item \href{#method-TPEa-initMessage}{\code{TPEa$initMessage()}}
\item \href{#method-TPEa-set_name}{\code{TPEa$set_name()}}
\item \href{#method-TPEa-set_model}{\code{TPEa$set_model()}}
\item \href{#method-TPEa-set_var}{\code{TPEa$set_var()}}
\item \href{#method-TPEa-set_env}{\code{TPEa$set_env()}}
\item \href{#method-TPEa-set_grid}{\code{TPEa$set_grid()}}
\item \href{#method-TPEa-set_map}{\code{TPEa$set_map()}}
\item \href{#method-TPEa-set_obs}{\code{TPEa$set_obs()}}
\item \href{#method-TPEa-set_weather}{\code{TPEa$set_weather()}}
\item \href{#method-TPEa-get_gridNames}{\code{TPEa$get_gridNames()}}
\item \href{#method-TPEa-get_varNames}{\code{TPEa$get_varNames()}}
\item \href{#method-TPEa-get_genotypes}{\code{TPEa$get_genotypes()}}
\item \href{#method-TPEa-get_envNames}{\code{TPEa$get_envNames()}}
\item \href{#method-TPEa-get_mapNames}{\code{TPEa$get_mapNames()}}
\item \href{#method-TPEa-get_gridid}{\code{TPEa$get_gridid()}}
\item \href{#method-TPEa-get_varid}{\code{TPEa$get_varid()}}
\item \href{#method-TPEa-get_envid}{\code{TPEa$get_envid()}}
\item \href{#method-TPEa-get_mapid}{\code{TPEa$get_mapid()}}
\item \href{#method-TPEa-createGrid}{\code{TPEa$createGrid()}}
\item \href{#method-TPEa-genClimate}{\code{TPEa$genClimate()}}
\item \href{#method-TPEa-runGridSim}{\code{TPEa$runGridSim()}}
\item \href{#method-TPEa-createMap}{\code{TPEa$createMap()}}
\item \href{#method-TPEa-plotMap}{\code{TPEa$plotMap()}}
\item \href{#method-TPEa-runEstimation}{\code{TPEa$runEstimation()}}
\item \href{#method-TPEa-print_grids}{\code{TPEa$print_grids()}}
\item \href{#method-TPEa-clone}{\code{TPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$new(
  name = "TPEa_1",
  model = "Samara",
  varieties = NA,
  environments = NA,
  genotypes = NA,
  vparameters = NA,
  eparameters = NA,
  parameters = NA,
  vep = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}

\item{\code{model}}{A character string with the name of the crop model used
for simulations}

\item{\code{varieties}}{A list of varieties names}

\item{\code{environments}}{A list of environments names}

\item{\code{genotypes}}{Optional. A list of alternate variety names}

\item{\code{vparameters}}{A vector or dataframe with variety specific parameters}

\item{\code{eparameters}}{A vector or dataframe with environment specific}

\item{\code{parameters}}{Optional. A vector or dataframe with all crop model
parameters (variety x environment)}

\item{\code{vep}}{Optional. If `parameters`is provided this argument
controls in what order parameters are given, either parameters for all
varieties of environment 1, then all for environment 2, etc. (by default,
TRUE) or parameters for all environments for variety 1, then for variety
2, etc. (FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$initMessage(vParamMissings, eParamMissings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vParamMissings}}{Name of varieties without parameters}

\item{\code{eParamMissings}}{Name of environments without parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPE analysis name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPE analysis name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set varieties, this is a list of TPEvar objects,
use with caution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New varieties}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_env"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_env}{}}}
\subsection{Method \code{set_env()}}{
Set environments, this is a list of TPEenv objects,
use with caution this will overwrite existing environments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_env(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New environments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_grid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_grid}{}}}
\subsection{Method \code{set_grid()}}{
Set grids, this is a list of grid objects,
use with caution this will overwrite existing grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_grid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New grids}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_map"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_map}{}}}
\subsection{Method \code{set_map()}}{
Set maps, this is a list of map objects,
use with caution, this will overwrite existing maps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_map(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New maps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_obs"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_obs}{}}}
\subsection{Method \code{set_obs()}}{
Set observation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_obs(varID = 1, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varID}}{A value of variety identifier (either index or name)}

\item{\code{val}}{A dataframe or list of dataframes with observations for each
of the `environments` (and eventually for each replicate)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_weather(envID = 1, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{envID}}{A value of environment identifier (either index or name)}

\item{\code{val}}{New weather}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_gridNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_gridNames}{}}}
\subsection{Method \code{get_gridNames()}}{
Get names of all grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_gridNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_varNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_varNames}{}}}
\subsection{Method \code{get_varNames()}}{
Get names of all varieties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_varNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_genotypes}{}}}
\subsection{Method \code{get_genotypes()}}{
Get names of all genotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_genotypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_envNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_envNames}{}}}
\subsection{Method \code{get_envNames()}}{
Get names of all environments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_envNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_mapNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_mapNames}{}}}
\subsection{Method \code{get_mapNames()}}{
Get names of all maps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_mapNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_gridid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_gridid}{}}}
\subsection{Method \code{get_gridid()}}{
Get grid id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_gridid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either grid name or a grid id (will return the grid id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_varid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_varid}{}}}
\subsection{Method \code{get_varid()}}{
Get var id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_varid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either var name or a var id (will return the var id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_envid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_envid}{}}}
\subsection{Method \code{get_envid()}}{
Get env id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_envid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either env name or a env id (will return the env id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-get_mapid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-get_mapid}{}}}
\subsection{Method \code{get_mapid()}}{
Get map id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$get_mapid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Either map name or a map id (will return the map id)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a simulation grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createGrid(
  name = "g1",
  varID = NA,
  res = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the grid}

\item{\code{varID}}{A variety identifier to use for simulation on the grid}

\item{\code{res}}{A numeric value of the resolution of the grid}

\item{\code{cols}}{A numeric value of the number of columns in the grid}

\item{\code{rows}}{A numeric value of the number of rows in the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the grid in
decimal degrees}

\item{\code{lat}}{A numeric value of the starting latitude of the grid in
decimal degrees}

\item{\code{multigrid}}{A boolean indicating if one grid should be created
for each genotype with the same cols, rows, lon and lat. If true, varID
will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$genClimate(
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  filesE = F,
  verbose = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{rcp}}{A character string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the  general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string with the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string with the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{filesE}}{Boolean. If weather files already exist}

\item{\code{verbose}}{Boolean. If messages about starting climate generation
should be shown}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runGridSim(
  gridID = NA,
  varID = NA,
  trait = "GrainYieldPopFin",
  year = 2015,
  soilData = soil,
  latlonData = lat_lon
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A variety identifier to use for
simulation on the grid. If NA, will use the variety attached to grid}

\item{\code{trait}}{A character string with the trait name for grid res matrix}

\item{\code{year}}{A numeric value with the year to run the simulation}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create raster map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createMap(name = "map1", gridID = NA, res = 150, bounds = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string defining the name of the map}

\item{\code{gridID}}{A grid identifier to use on the map}

\item{\code{res}}{A numeric value in sec of the resolution of world map to use
Options are c(1, 150, 900) for respectively 30sec, 2.5min, 15min}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$plotMap(mapID = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A map identifier to plot
(either index or name). By default will run on all maps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$runEstimation(
  varID = 1,
  envID = 1,
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varID}}{A value or list of values of variety identifier
(either index or name)}

\item{\code{envID}}{A value or list of values of environment identifier
(either index or name)}

\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment}

\item{\code{bounds}}{A matrix with lower (col1) and upper (col2) bounds for
each of the parameters in `paramnames` (rows)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-print_grids"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-print_grids}{}}}
\subsection{Method \code{print_grids()}}{
Print grid information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$print_grids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
