% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tpe.R
\name{TPEa}
\alias{TPEa}
\title{#' R6 Class Representing a TPE analysis}
\description{
TODO
}
\details{
TODO
}
\examples{

## ------------------------------------------------
## Method `TPEa$set_model`
## ------------------------------------------------

TPE_analysis <- TPEa("TPEa1", "STICS")
TPE_analysis$model
TPE_analysis$set_model("Samara")
TPE_analysis$model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{model}}{Name of the crop model used for simulations}

\item{\code{varieties}}{List of varieties names used for TPE analysis}

\item{\code{environments}}{List of environment names used for TPE analysis}

\item{\code{genotypes}}{List of alternate varieties names used for TPE analysis}

\item{\code{latStart}}{Starting latitude for the grid (upper left corner)}

\item{\code{lonStart}}{Starting longitude for the grid (upper left corner)}

\item{\code{grid}}{Simulation grid}

\item{\code{parameters}}{A vector or dataframe with all parameters used for
simulation with the corresponding `model`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEa-new}{\code{TPEa$new()}}
\item \href{#method-TPEa-set_model}{\code{TPEa$set_model()}}
\item \href{#method-TPEa-initMessage}{\code{TPEa$initMessage()}}
\item \href{#method-TPEa-createGrid}{\code{TPEa$createGrid()}}
\item \href{#method-TPEa-genClim}{\code{TPEa$genClim()}}
\item \href{#method-TPEa-clone}{\code{TPEa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$new(
  name = "TPEa_1",
  model = "Samara",
  varieties = NA,
  environments = NA,
  latStart = NA,
  lonStart = NA,
  genotypes = NA,
  parameters = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{model}}{Name of the crop model used for simulations}

\item{\code{varieties}}{List of varieties}

\item{\code{environments}}{List of environments}

\item{\code{latStart}}{Starting latitude for grid}

\item{\code{lonStart}}{Starting longitude for grid}

\item{\code{genotypes}}{Optional. List of alternate variety names}

\item{\code{parameters}}{A vector or dataframe with all crop model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEa` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-set_model}{}}}
\subsection{Method \code{set_model()}}{
Change crop model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{TPE_analysis <- TPEa("TPEa1", "STICS")
TPE_analysis$model
TPE_analysis$set_model("Samara")
TPE_analysis$model
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a simulation grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$createGrid(res = 5, cols = NA, rows = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{Resolution of grid, in km}

\item{\code{cols}}{Number of columns in the grid}

\item{\code{rows}}{Number of rows in the grid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-genClim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-genClim}{}}}
\subsection{Method \code{genClim()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$genClim(
  rcp = "rcp26",
  year = 2015,
  yearNb = 99,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
