% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_analysis.R
\name{TPEase}
\alias{TPEase}
\title{#' R6 Class Representing a TPEase analysis}
\description{
The TPEase `analysis` object is an environment containing all
objects used for parameter estimation and TPE analysis
(varieties, environments, grids, maps, etc.)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEase analysis}

\item{\code{model}}{A character string identifier of the used crop model
(only "Samara" is supported for now)}

\item{\code{varieties}}{A list of varieties objects}

\item{\code{TPEanalysis}}{A list of TPE analysis objects}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEase-new}{\code{TPEase$new()}}
\item \href{#method-TPEase-initMessage}{\code{TPEase$initMessage()}}
\item \href{#method-TPEase-set_name}{\code{TPEase$set_name()}}
\item \href{#method-TPEase-set_model}{\code{TPEase$set_model()}}
\item \href{#method-TPEase-set_var}{\code{TPEase$set_var()}}
\item \href{#method-TPEase-get_varNames}{\code{TPEase$get_varNames()}}
\item \href{#method-TPEase-get_genotypes}{\code{TPEase$get_genotypes()}}
\item \href{#method-TPEase-get_varid}{\code{TPEase$get_varid()}}
\item \href{#method-TPEase-createTPE}{\code{TPEase$createTPE()}}
\item \href{#method-TPEase-createGrid}{\code{TPEase$createGrid()}}
\item \href{#method-TPEase-genClimate}{\code{TPEase$genClimate()}}
\item \href{#method-TPEase-runGridSim}{\code{TPEase$runGridSim()}}
\item \href{#method-TPEase-createMap}{\code{TPEase$createMap()}}
\item \href{#method-TPEase-runClustering}{\code{TPEase$runClustering()}}
\item \href{#method-TPEase-plotMap}{\code{TPEase$plotMap()}}
\item \href{#method-TPEase-addToPlot}{\code{TPEase$addToPlot()}}
\item \href{#method-TPEase-print_maps}{\code{TPEase$print_maps()}}
\item \href{#method-TPEase-runEstimation}{\code{TPEase$runEstimation()}}
\item \href{#method-TPEase-clone}{\code{TPEase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPEase analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$new(
  name = "TPEase_analysis",
  model = "Samara",
  varieties = NA,
  genotypes = NA,
  vparameters = NA,
  environments = NA,
  eparameters = NA,
  eName = "estim1",
  weathers = NA,
  observations = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEase analysis}

\item{\code{model}}{A character string identifier of the crop model used}

\item{\code{varieties}}{A list of varieties names}

\item{\code{genotypes}}{Optional. A list of alternate variety names}

\item{\code{vparameters}}{A vector or dataframe with variety specific parameters}

\item{\code{environments}}{A list of environments names}

\item{\code{eparameters}}{A vector or dataframe with environment specific
parameters}

\item{\code{eName}}{A character string identifier of the initial estimation
object}

\item{\code{weathers}}{A list (each entry is an environment) of dataframes
containing weather data}

\item{\code{observations}}{A list (each entry is a variety) of lists
(each entry is an environment) of dataframes with observations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEase` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$initMessage(vParamMissings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vParamMissings}}{Name of varieties without parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set TPEase analysis name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New TPEase analysis name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_model"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_model}{}}}
\subsection{Method \code{set_model()}}{
Set model name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_model(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New model name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set varieties, this is a list of TPEvar objects,
use with caution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New varieties}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_varNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_varNames}{}}}
\subsection{Method \code{get_varNames()}}{
Get names of all varieties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_varNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_genotypes"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_genotypes}{}}}
\subsection{Method \code{get_genotypes()}}{
Get names of all genotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_genotypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-get_varid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-get_varid}{}}}
\subsection{Method \code{get_varid()}}{
Get variety id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$get_varid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A variety identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createTPE"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createTPE}{}}}
\subsection{Method \code{createTPE()}}{
Create a TPE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createTPE(name = "TPEa_1")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string of TPE name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createGrid}{}}}
\subsection{Method \code{createGrid()}}{
Create a grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createGrid(
  tpeID = 1,
  name = "g1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  cols = 5,
  rows = 5,
  lon = NA,
  lat = NA,
  multigrid = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{name}}{A character string identifier of the grid}

\item{\code{varID}}{A variety identifier to use for simulation on the grid}

\item{\code{latres}}{A numeric value of the latitude resolution of the grid}

\item{\code{lonres}}{A numeric value of the longitude resolution of the grid}

\item{\code{cols}}{A numeric value of the number of columns in the grid}

\item{\code{rows}}{A numeric value of the number of rows in the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the grid in
decimal degrees}

\item{\code{lat}}{A numeric value of the starting latitude of the grid in
decimal degrees}

\item{\code{multigrid}}{A boolean indicating if one grid should be created
for each genotype with the same cols, rows, lon and lat. If true, varID
will be ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$genClimate(
  tpeID = 1,
  gridID = NA,
  rcp = "rcp26",
  year = 2014,
  yearNb = 1,
  modelNb = "00000000000000000",
  path = NA,
  pathCLI = NA,
  seed = NA,
  filesE = F,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{rcp}}{A character string of the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A numeric value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A numeric value of the number of years to simulate}

\item{\code{modelNb}}{A character string of the general circulation model
identifier to use, see \code{generateClimate}}

\item{\code{path}}{A character string of the path to the marksim standalone.
For the moment, this path can not contain spaces}

\item{\code{pathCLI}}{Optional. A character string of the path to CLI folder.
If no value is provided, the CLI folder will be considered in the same
folder as the marksim standalone. For the moment, this path can not
contain spaces}

\item{\code{seed}}{A numeric value of the seed to use to generate climate}

\item{\code{filesE}}{Boolean. If weather files already exist and should be used
(existing weather files should be located in a "weathers" subfolder of
the marksim path)}

\item{\code{verbose}}{Boolean. If messages about starting climate generation
should be shown
TODO: check if verbose is used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation on one or several grids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runGridSim(
  tpeID = 1,
  gridID = NA,
  varID = NA,
  soilData = soil,
  latlonData = lat_lon,
  cumulP = cumul,
  traitList = NULL,
  savePath = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{gridID}}{Optional. A vector of grid identifiers
(either index or name). By default will run all grids}

\item{\code{varID}}{A variety identifier to use for
simulation on the grid. If NA, will use the variety attached to grid}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.
TODO: implement generic methods for soil, latlon, cumul}

\item{\code{traitList}}{Optional. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the
maximum values for each year of the selected traits}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-createMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-createMap}{}}}
\subsection{Method \code{createMap()}}{
Create raster map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$createMap(tpeID = 1, name = "map1", bounds = NA, res = 150)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{name}}{A character string defining the name of the map}

\item{\code{bounds}}{Optional. A vector of four numeric values as decimal degree
of north, east, south, west bounds to crop map
(by default, with no bounds, will display the map of the world)}

\item{\code{res}}{Not used at the moment. A numeric value in sec of the
resolution of world map to use. Options are c(1, 150, 900)
for respectively 30sec, 2.5min, 15min
TODO: remove resolution of map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runClustering"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runClustering}{}}}
\subsection{Method \code{runClustering()}}{
Run TPE clustering by performing Principle Component
Analysis (PCA) and Hierarchical Clustering on Principle Component (HCPC)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runClustering(
  tpeID = 1,
  mapID = 1,
  traitList = c("GrainYieldPop"),
  nbDim = 5,
  nbClust = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{mapID}}{A value or list of values of map identifiers
(either index or names) on which to perform clustering}

\item{\code{traitList}}{Vector of variable names to be used for PCA}

\item{\code{nbDim}}{Integer of number of dimensions to use for PCA}

\item{\code{nbClust}}{Integer of number of clusters to use for HCPC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$plotMap(tpeID = 1, mapID = 1, trait = "cluster", isFactor = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A TPE identifier on which to create grid}

\item{\code{mapID}}{Id of map to plot}

\item{\code{trait}}{A character string identifier of the data to plot,
by default will plot the cluster computed by the runClustering function
of the TPE analysis object}

\item{\code{isFactor}}{A boolean indicating if the trait should be considered
as a factor for plotting (cluster is a factor)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-addToPlot"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-addToPlot}{}}}
\subsection{Method \code{addToPlot()}}{
Add ggplot2 objects to existing plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$addToPlot(tpeID = 1, mapID = 1, plotID = 1, plotAdd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A numeric value identifier of the TPE}

\item{\code{mapID}}{A numeric value identifier of the map}

\item{\code{plotID}}{A numeric value identifier of the plot}

\item{\code{plotAdd}}{A list of ggplot2 objects to pass to the plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-print_maps"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-print_maps}{}}}
\subsection{Method \code{print_maps()}}{
Print a given map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$print_maps(tpeID = 1, mapID = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpeID}}{A numeric value identifier of the TPE}

\item{\code{mapID}}{A numeric value identifier of the map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$runEstimation(
  varID = 1,
  estimID = 1,
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varID}}{A value or list of values of variety identifier
(either index or name)}

\item{\code{estimID}}{A value of estimation identifier (either index or name)}

\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment
TODO: implement weigh_fn}

\item{\code{bounds}}{A matrix with lower (row1) and upper (row2) bounds for
each of the parameters in `paramnames` (cols)}

\item{\code{...}}{Additional parameters to be passed to DEoptim.control}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
