% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_var.R
\name{TPEaseVar}
\alias{TPEaseVar}
\title{R6 Class representing a TPEase Variety}
\usage{
Varieties are automatically created when the parent \code{TPEase} is
initialized
}
\description{
Variety object containing all information about a crop variety
}
\details{
The \code{TPEaseVar} object is the first brick built upon the \code{TPEase}.
It contains basic information about a crop variety as well as a list of
\code{VarEstim} objects (parameter estimation)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The \code{TPEaseVar} identifier}

\item{\code{alt}}{An alternate name for the variety}

\item{\code{parent}}{The \code{TPEase} object parent of the \code{TPEaseVar}}

\item{\code{parameters}}{A \code{data.frame} with all parameters used
for simulation}

\item{\code{estimations}}{A list of \code{VarEstim} objects}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEaseVar-new}{\code{TPEaseVar$new()}}
\item \href{#method-TPEaseVar-get_estimNames}{\code{TPEaseVar$get_estimNames()}}
\item \href{#method-TPEaseVar-get_estimid}{\code{TPEaseVar$get_estimid()}}
\item \href{#method-TPEaseVar-set_param}{\code{TPEaseVar$set_param()}}
\item \href{#method-TPEaseVar-updateParam}{\code{TPEaseVar$updateParam()}}
\item \href{#method-TPEaseVar-createEstim}{\code{TPEaseVar$createEstim()}}
\item \href{#method-TPEaseVar-runEstimation}{\code{TPEaseVar$runEstimation()}}
\item \href{#method-TPEaseVar-clone}{\code{TPEaseVar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TPEaseVar} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$new(
  name = "v1",
  alt = "G1",
  parameters = NA,
  eName = NA,
  environments = NA,
  eparam = NA,
  weathers = NA,
  observations = NA,
  parent = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of the TPEaseVar}

\item{\code{alt}}{A \code{character} string alternate identifier of the variety}

\item{\code{parameters}}{A \code{data.frame} with variety parameters}

\item{\code{eName}}{Optionnal. A \code{character} string identifier of initial
\code{VarEstim} object. By default, will not create a \code{VarEstim}
object}

\item{\code{environments}}{A \code{vector} of environments names}

\item{\code{eparam}}{A \code{data.frame} with environment parameters}

\item{\code{weathers}}{A \code{list} of \code{data.frames} of weathers
(each entry is for one environment)}

\item{\code{observations}}{A \code{list} of \code{data.frames} of observations
(each entry is for one environment)}

\item{\code{parent}}{A \code{TPEase} object parent of the \code{TPEaseVar}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TPEaseVar} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-get_estimNames"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-get_estimNames}{}}}
\subsection{Method \code{get_estimNames()}}{
Get \code{VarEstim} names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$get_estimNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-get_estimid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-get_estimid}{}}}
\subsection{Method \code{get_estimid()}}{
Get \code{VarEstim} id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$get_estimid(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{VarEstim} identifier (either index or name)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-set_param"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-set_param}{}}}
\subsection{Method \code{set_param()}}{
Set \code{TPEaseVar} parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$set_param(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{data.frame} with variety parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-updateParam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-updateParam}{}}}
\subsection{Method \code{updateParam()}}{
Set \code{TPEaseVar} estimated parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$updateParam(val, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{vector} with parameter values}

\item{\code{names}}{A \code{vector} with estimated parameter names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-createEstim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-createEstim}{}}}
\subsection{Method \code{createEstim()}}{
Create a \code{VarEstim} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$createEstim(
  name,
  parent,
  environments,
  eparam,
  weathers,
  observations
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of the \code{VarEstim}}

\item{\code{parent}}{A \code{TPEaseVar} object parent of the \code{VarEstim}}

\item{\code{environments}}{A \code{vector} of environments identifiers}

\item{\code{eparam}}{A \code{data.frame} with parameters of environments}

\item{\code{weathers}}{A \code{data.frame} with weather data of environments}

\item{\code{observations}}{A \code{data.frame} with observations of environments
for the variety \code{TPEaseVar}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$runEstimation(
  estimID = 1,
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimID}}{A \code{VarEstim} identifier (either index or name)}

\item{\code{maxiter}}{A \code{numeric} value of the maximum number of iteration
for \code{DEoptim}}

\item{\code{paramnames}}{A \code{vector} of parameter names to be estimated}

\item{\code{metric}}{A \code{character} string with the name of the metric
to use for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A \code{function} to compute fitness,
see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment
TODO: implement weigh_fn}

\item{\code{bounds}}{A \code{matrix} or \code{data.frame}
with lower (col1) and upper (col2) bounds for each of the parameters in
\code{paramnames} (rows)}

\item{\code{args}}{Additional parameters to be passed to \code{DEoptim.control}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEaseVar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEaseVar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEaseVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
