% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_cm.R
\name{TPEcm}
\alias{TPEcm}
\title{#' R6 Class Representing a crop managment}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{soilParam}}{soil parameters of grid point}

\item{\code{dateParam}}{sowing date of grid point}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEcm-new}{\code{TPEcm$new()}}
\item \href{#method-TPEcm-clone}{\code{TPEcm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-new}{}}}
\subsection{Method \code{new()}}{
Create a new variety object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$new(name = "e1", parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEenv` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEcm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEcm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEcm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
