% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEGrid}
\alias{TPEGrid}
\title{R6 Class representing a TPEGrid grid}
\usage{
res <- TPEase$new(...)
res$createTpe("TPE1")
res$createGrid(tpeID=1, "Grid1", varID, latres, lonres, cols, rows, lon, lat,
               multigrid)
}
\description{
Grid object summarizing information and simulations on the different
\code{GridPoints}
}
\details{
The \code{TPEGrid} object contains all information of the grid as well as a
collection of \code{GridPoints}. The object summarizes simulations in all
environments (\code{GridPoints}) for a given variety (\code{TPEaseVar})
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The identifier of the \code{TPEGrid}}

\item{\code{latRes}}{Latitude resolution of the grid}

\item{\code{lonRes}}{Longitude resolution of the grid}

\item{\code{lonStart}}{Starting longitude of the grid (in decimal degrees)}

\item{\code{latStart}}{Starting latitude of the grid (in decimal degrees)}

\item{\code{gridPoints}}{A \code{matrix} of \code{GridPoint} objects}

\item{\code{gridRes}}{Summary of simulation results at the grid level}

\item{\code{parent}}{The \code{TPEaseTPE} parent object of \code{TPEGrid}}

\item{\code{variety}}{A \code{TPEaseVar} to which this grid is linked to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEGrid-new}{\code{TPEGrid$new()}}
\item \href{#method-TPEGrid-initMessage}{\code{TPEGrid$initMessage()}}
\item \href{#method-TPEGrid-set_var}{\code{TPEGrid$set_var()}}
\item \href{#method-TPEGrid-set_name}{\code{TPEGrid$set_name()}}
\item \href{#method-TPEGrid-set_gridres}{\code{TPEGrid$set_gridres()}}
\item \href{#method-TPEGrid-get_width}{\code{TPEGrid$get_width()}}
\item \href{#method-TPEGrid-get_length}{\code{TPEGrid$get_length()}}
\item \href{#method-TPEGrid-get_results}{\code{TPEGrid$get_results()}}
\item \href{#method-TPEGrid-add_gridpointResult}{\code{TPEGrid$add_gridpointResult()}}
\item \href{#method-TPEGrid-populateGrid}{\code{TPEGrid$populateGrid()}}
\item \href{#method-TPEGrid-genClimate}{\code{TPEGrid$genClimate()}}
\item \href{#method-TPEGrid-runGridSim}{\code{TPEGrid$runGridSim()}}
\item \href{#method-TPEGrid-clone}{\code{TPEGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TPEGrid} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$new(
  name = "grid1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  width = 5,
  length = 5,
  lon = NA,
  lat = NA,
  parent = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A \code{character} string identifier of the \code{TPEGrid}}

\item{\code{varID}}{A \code{TPEaseVar} identifier to link to this grid}

\item{\code{latres}}{A \code{numeric} value of the latitude resolution of the
grid (in decimal degrees)}

\item{\code{lonres}}{A \code{numeric} value of the longitude resolution of the
grid (in decimal degrees)}

\item{\code{width}}{A \code{numeric} value of the Width of the grid}

\item{\code{length}}{A \code{numeric} value of the length of the grid}

\item{\code{lon}}{A \code{numeric} value of the starting longitude of the
grid (upper left corner)}

\item{\code{lat}}{A \code{numeric} value of the starting latitude of the grid
(upper left corner)}

\item{\code{parent}}{The \code{TPEaseTPE} parent object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEGrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of \code{TPEGrid} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set variety
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{\code{TPEaseVar} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set \code{TPEGrid} name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New \code{TPEGrid} name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_gridres"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_gridres}{}}}
\subsection{Method \code{set_gridres()}}{
Set grid result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_gridres(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{data.frame} of simulation results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_width}{}}}
\subsection{Method \code{get_width()}}{
Get grid width
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_length"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_length}{}}}
\subsection{Method \code{get_length()}}{
Get grid length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get grid simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_results(traitList = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{A \code{vector} of variable names to extract
from simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-add_gridpointResult"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-add_gridpointResult}{}}}
\subsection{Method \code{add_gridpointResult()}}{
Add result of a \code{GridPoint} to \code{gridRes}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$add_gridpointResult(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A \code{data.frame} of \code{GridPoint} results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-populateGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-populateGrid}{}}}
\subsection{Method \code{populateGrid()}}{
Create \code{GridPoint} for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$populateGrid(lat = NA, lon = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{A \code{numeric} value of the starting latitude of the grid
(upper left corner)}

\item{\code{lon}}{A \code{numeric} value of the starting longitude of the grid
(upper left corner)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$genClimate(
  rcp,
  year,
  yearNb,
  modelNb,
  path,
  pathCLI,
  filesE,
  verbose,
  seed
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{A \code{character} string with the name of the Representative
Concentration Pathway to use. One of the following options
c("rcp26","rcp45","rcp60","rcp85")}

\item{\code{year}}{A \code{numeric} value of the year to simulate climate
(this can include years from 2013 to 2099)}

\item{\code{yearNb}}{A \code{numeric} value of the number of years to simulate}

\item{\code{modelNb}}{A \code{character} string of the  general circulation
model identifier to use, see \code{generateClimate}}

\item{\code{path}}{A \code{character} string with the path to the marksim
standalone. This path can not contain spaces}

\item{\code{pathCLI}}{Optional. A \code{character} string with the path to
CLI folder. If no value is provided, the CLI folder will be considered
in the same folder as the marksim standalone. This path can not
contain spaces}

\item{\code{filesE}}{A \code{boolean} indicating if weather files already exist}

\item{\code{verbose}}{A \code{boolean} indicating ff messages about starting
climate generation should be shown}

\item{\code{seed}}{An \code{integer} number to use as seed for marksim weather
generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$runGridSim(soilData, latlonData, cumulP, traitList, savePath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. A \code{vector} of trait names to extract
from simulations. This will delete the simulations and only keep the
mean of the yearly maximum for each trait}

\item{\code{savePath}}{Optional. A \code{character} string of the path where
to save simulation files. If \code{NULL} (default),
will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
