% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEGrid}
\alias{TPEGrid}
\title{R6 Class representing a TPEGrid}
\description{
The `TPEGrid` object contains all information of the grid as well as a
collection of GridPoints
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{latRes}}{A numeric value of latitude resolution of the grid}

\item{\code{lonRes}}{A numeric value of longitude resolution of the grid}

\item{\code{lonStart}}{A numeric value of the starting longitude
of the grid}

\item{\code{latStart}}{A numeric value of the starting latitude
of the grid}

\item{\code{gridPoints}}{A matrix of GridPoint objects}

\item{\code{gridRes}}{A dataframe with simulation results}

\item{\code{parent}}{A dataframe object parent of the TPEGrid}

\item{\code{variety}}{A TPEaseVar to which this grid is linked to}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEGrid-new}{\code{TPEGrid$new()}}
\item \href{#method-TPEGrid-initMessage}{\code{TPEGrid$initMessage()}}
\item \href{#method-TPEGrid-set_var}{\code{TPEGrid$set_var()}}
\item \href{#method-TPEGrid-set_name}{\code{TPEGrid$set_name()}}
\item \href{#method-TPEGrid-set_gridres}{\code{TPEGrid$set_gridres()}}
\item \href{#method-TPEGrid-get_width}{\code{TPEGrid$get_width()}}
\item \href{#method-TPEGrid-get_length}{\code{TPEGrid$get_length()}}
\item \href{#method-TPEGrid-get_results}{\code{TPEGrid$get_results()}}
\item \href{#method-TPEGrid-add_gridpointResult}{\code{TPEGrid$add_gridpointResult()}}
\item \href{#method-TPEGrid-populateGrid}{\code{TPEGrid$populateGrid()}}
\item \href{#method-TPEGrid-genClimate}{\code{TPEGrid$genClimate()}}
\item \href{#method-TPEGrid-runGridSim}{\code{TPEGrid$runGridSim()}}
\item \href{#method-TPEGrid-clone}{\code{TPEGrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPEGrid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$new(
  name = "grid1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  width = 5,
  length = 5,
  lon = NA,
  lat = NA,
  parent = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPEGrid}

\item{\code{varID}}{A TPEaseVar identifier to which this grid is linked}

\item{\code{latres}}{A numeric value of the latitude resolution of the grid
(in decimal degrees)}

\item{\code{lonres}}{A numeric value of the longitude resolution of the grid
(in decimal degrees)}

\item{\code{width}}{A numeric value of the Width of the grid}

\item{\code{length}}{A numeric value of the length of the grid}

\item{\code{lon}}{A numeric value of the starting longitude of the
grid (upper left corner)}

\item{\code{lat}}{A numeric value of the starting latitude of the grid
(upper left corner)}

\item{\code{parent}}{A TPEaseVar to which this grid is linked to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEGrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of Grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set variety
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Variety}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set grid name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Variety}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-set_gridres"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-set_gridres}{}}}
\subsection{Method \code{set_gridres()}}{
Set grid result dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$set_gridres(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_width}{}}}
\subsection{Method \code{get_width()}}{
Get grid width
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_length"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_length}{}}}
\subsection{Method \code{get_length()}}{
Get grid length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get grid simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$get_results(traitList = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traitList}}{Vector of variable names to extract for results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-add_gridpointResult"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-add_gridpointResult}{}}}
\subsection{Method \code{add_gridpointResult()}}{
Add result of gridPoint to gridRes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$add_gridpointResult(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe of gridPoint results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-populateGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-populateGrid}{}}}
\subsection{Method \code{populateGrid()}}{
Create gridpoint for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$populateGrid(lat = NA, lon = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Starting latitude of the grid (upper left corner)}

\item{\code{lon}}{Starting longitude of the grid (upper left corner)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$genClimate(
  rcp,
  year,
  yearNb,
  modelNb,
  path,
  pathCLI,
  filesE,
  verbose,
  seed
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}

\item{\code{filesE}}{Boolean. If weather files already exist}

\item{\code{verbose}}{Boolean. If messages about completing climate generation
should be shown}

\item{\code{seed}}{Integer number to use as seed for marksim weather generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$runGridSim(soilData, latlonData, cumulP, traitList, savePath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{cumulP}}{Tmp for Adam et al.}

\item{\code{traitList}}{Optionnal. Vector of trait names to extract from
simulations. This will delete the simulations and only keep the mean of
the yearly maximum for each trait}

\item{\code{savePath}}{Optional. A character string of the path where to save
simulation files. If NULL (default), will not save simulations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
