% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEgrid}
\alias{TPEgrid}
\title{#' R6 Class Representing a simulation grid}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{length}}{Length of the grid}

\item{\code{width}}{Width of the grid}

\item{\code{res}}{Resolution of the grid}

\item{\code{lonStart}}{Starting longitude of the grid}

\item{\code{latStart}}{Starting latitude of the grid}

\item{\code{gridPoints}}{Collection of each point on the grid}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEgrid-new}{\code{TPEgrid$new()}}
\item \href{#method-TPEgrid-clone}{\code{TPEgrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$new(name = "grid1", res = 5, width = 5, length = 5, lon = NA, lat = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{res}}{Resolution of the grid}

\item{\code{width}}{Width of the grid}

\item{\code{length}}{Length of the grid}

\item{\code{lon}}{Optional. Starting longitude of the grid (upper left corner)}

\item{\code{lat}}{Optional. Starting latitude of the grid (upper left corner)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
