% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEgrid}
\alias{TPEgrid}
\title{#' R6 Class Representing a simulation grid}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{length}}{Length of the grid}

\item{\code{width}}{Width of the grid}

\item{\code{res}}{Resolution of the grid}

\item{\code{lonStart}}{Starting longitude of the grid}

\item{\code{latStart}}{Starting latitude of the grid}

\item{\code{gridPoints}}{Collection of each point on the grid}

\item{\code{resGrid}}{Matrix with value of simulated trait for each grid point}

\item{\code{parent}}{TPE analysis parent}

\item{\code{plots}}{A list of map plots generated by the plotMap() function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEgrid-new}{\code{TPEgrid$new()}}
\item \href{#method-TPEgrid-set_resGrid}{\code{TPEgrid$set_resGrid()}}
\item \href{#method-TPEgrid-genClimate}{\code{TPEgrid$genClimate()}}
\item \href{#method-TPEgrid-runGridSim}{\code{TPEgrid$runGridSim()}}
\item \href{#method-TPEgrid-plotMap}{\code{TPEgrid$plotMap()}}
\item \href{#method-TPEgrid-clone}{\code{TPEgrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$new(
  name = "grid1",
  res = 5,
  width = 5,
  length = 5,
  lon = NA,
  lat = NA,
  parent = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{res}}{Resolution of the grid}

\item{\code{width}}{Width of the grid}

\item{\code{length}}{Length of the grid}

\item{\code{lon}}{Optional. Starting longitude of the grid (upper left corner)}

\item{\code{lat}}{Optional. Starting latitude of the grid (upper left corner)}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-set_resGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-set_resGrid}{}}}
\subsection{Method \code{set_resGrid()}}{
Set simulation result for given grid point in result matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$set_resGrid(i, j, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Row position of grid point}

\item{\code{j}}{Col position of grid point}

\item{\code{val}}{Value of simulation result}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$genClimate(rcp, year, yearNb, modelNb, path, pathCLI)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$runGridSim(trait, year, soilData, latlonData, param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trait}}{Trait name for grid res matrix}

\item{\code{year}}{Year to run the simulation}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}

\item{\code{param}}{Parameters to use for simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-plotMap"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-plotMap}{}}}
\subsection{Method \code{plotMap()}}{
Create plot on map based on grid simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$plotMap(mapID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapID}}{A numeric value of the index of the map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
