% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEgrid}
\alias{TPEgrid}
\title{#' R6 Class Representing a simulation grid point}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid point}

\item{\code{lat}}{Latitude of the TPE analysis}

\item{\code{lon}}{Longitude of the TPE analysis}

\item{\code{weather}}{Samara climate data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEgrid-new}{\code{TPEgrid$new()}}
\item \href{#method-TPEgrid-set_weather}{\code{TPEgrid$set_weather()}}
\item \href{#method-TPEgrid-clone}{\code{TPEgrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$new(name = "c11", lat = 11.18, lon = -4.3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the TPE analysis}

\item{\code{lat}}{Latitude of grid point}

\item{\code{lon}}{Longitude of grid point}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Change weather of grid point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$set_weather(weather)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weather}}{Weather dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
