% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_grid.R
\name{TPEgrid}
\alias{TPEgrid}
\title{#' R6 Class Representing a simulation grid}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{latRes}}{Latitude resolution of the grid}

\item{\code{lonRes}}{Longitude resolution of the grid}

\item{\code{lonStart}}{Starting longitude of the grid}

\item{\code{latStart}}{Starting latitude of the grid}

\item{\code{gridPoints}}{Collection of each point on the grid}

\item{\code{gridRes}}{Result dataframe (constructed during map clustering)}

\item{\code{parent}}{TPE analysis parent}

\item{\code{variety}}{Variety to which this grid is linked to}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEgrid-new}{\code{TPEgrid$new()}}
\item \href{#method-TPEgrid-initMessage}{\code{TPEgrid$initMessage()}}
\item \href{#method-TPEgrid-set_var}{\code{TPEgrid$set_var()}}
\item \href{#method-TPEgrid-set_name}{\code{TPEgrid$set_name()}}
\item \href{#method-TPEgrid-set_gridres}{\code{TPEgrid$set_gridres()}}
\item \href{#method-TPEgrid-get_width}{\code{TPEgrid$get_width()}}
\item \href{#method-TPEgrid-get_length}{\code{TPEgrid$get_length()}}
\item \href{#method-TPEgrid-get_results}{\code{TPEgrid$get_results()}}
\item \href{#method-TPEgrid-populateGrid}{\code{TPEgrid$populateGrid()}}
\item \href{#method-TPEgrid-genClimate}{\code{TPEgrid$genClimate()}}
\item \href{#method-TPEgrid-runGridSim}{\code{TPEgrid$runGridSim()}}
\item \href{#method-TPEgrid-clone}{\code{TPEgrid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$new(
  name = "grid1",
  varID = NA,
  latres = 0.35,
  lonres = 0.5,
  width = 5,
  length = 5,
  lon = NA,
  lat = NA,
  parent = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Identifier of the grid}

\item{\code{varID}}{Identifier of the variety to which this grid is linked}

\item{\code{latres}}{Latitude resolution of the grid (in decimal degrees)}

\item{\code{lonres}}{Longitude resolution of the grid (in decimal degrees)}

\item{\code{width}}{Width of the grid}

\item{\code{length}}{Length of the grid}

\item{\code{lon}}{Optional. Starting longitude of the grid (upper left corner)}

\item{\code{lat}}{Optional. Starting latitude of the grid (upper left corner)}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEgrid` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-initMessage"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-initMessage}{}}}
\subsection{Method \code{initMessage()}}{
Confirm creation of Grid object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$initMessage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-set_var"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-set_var}{}}}
\subsection{Method \code{set_var()}}{
Set variety
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$set_var(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Variety}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set grid name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$set_name(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Variety}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-set_gridres"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-set_gridres}{}}}
\subsection{Method \code{set_gridres()}}{
Set grid result dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$set_gridres(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-get_width}{}}}
\subsection{Method \code{get_width()}}{
Get grid width
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-get_length"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-get_length}{}}}
\subsection{Method \code{get_length()}}{
Get grid length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$get_length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get grid simulation results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$get_results(varList = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varList}}{Vector of variable names to extract for results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-populateGrid"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-populateGrid}{}}}
\subsection{Method \code{populateGrid()}}{
Create gridpoint for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$populateGrid(lat = NA, lon = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Starting latitude of the grid (upper left corner)}

\item{\code{lon}}{Starting longitude of the grid (upper left corner)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate data for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$genClimate(rcp, year, yearNb, modelNb, path, pathCLI, filesE, verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}

\item{\code{filesE}}{Boolean. If weather files already exist}

\item{\code{verbose}}{Boolean. If messages about completing climate generation
should be shown}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-runGridSim"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-runGridSim}{}}}
\subsection{Method \code{runGridSim()}}{
Run simulation for each point of the grid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$runGridSim(year, soilData, latlonData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Year to run the simulation}

\item{\code{soilData}}{Tmp for Adam et al.}

\item{\code{latlonData}}{Tmp for Adam et al.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEgrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEgrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEgrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
