% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_map.R
\name{TPEmap}
\alias{TPEmap}
\title{#' R6 Class Representing a TPE map}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the map}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEmap-new}{\code{TPEmap$new()}}
\item \href{#method-TPEmap-clone}{\code{TPEmap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE analysis object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$new(name = "map1")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE analysis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEmap` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEmap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEmap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEmap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
