% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_var.R
\name{TPEvar}
\alias{TPEvar}
\title{#' R6 Class Representing a TPE analysis}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{alt}}{A character string alternate adientifier of the variety}

\item{\code{parent}}{TPE analysis parent}

\item{\code{parameters}}{A dataframe with all parameters used for simulation}

\item{\code{estimParam}}{A vector containing the values of estimated parameters}

\item{\code{observations}}{A dataframe or list of dataframe with observations
for each of the `environments`. If several replicates, a list of lists
(environment x replicate)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEvar-new}{\code{TPEvar$new()}}
\item \href{#method-TPEvar-set_obs}{\code{TPEvar$set_obs()}}
\item \href{#method-TPEvar-set_param}{\code{TPEvar$set_param()}}
\item \href{#method-TPEvar-set_eparam}{\code{TPEvar$set_eparam()}}
\item \href{#method-TPEvar-update_param}{\code{TPEvar$update_param()}}
\item \href{#method-TPEvar-runEstimation}{\code{TPEvar$runEstimation()}}
\item \href{#method-TPEvar-clone}{\code{TPEvar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-new}{}}}
\subsection{Method \code{new()}}{
Create a new variety object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$new(name = "v1", alt = "G1", parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the variety}

\item{\code{alt}}{A character string alternate identifier of the variety}

\item{\code{parent}}{TPE analysis parent}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEvar` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-set_obs"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-set_obs}{}}}
\subsection{Method \code{set_obs()}}{
Set observation data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$set_obs(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe or list of dataframes with observations for each
of the `environments` (and eventually for each replicate)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-set_param"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-set_param}{}}}
\subsection{Method \code{set_param()}}{
Set parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$set_param(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe with all crop model parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-set_eparam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-set_eparam}{}}}
\subsection{Method \code{set_eparam()}}{
Set estimated parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$set_eparam(val, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A vector with estimated parameter values}

\item{\code{names}}{A vector with estimated parameter names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-update_param"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-update_param}{}}}
\subsection{Method \code{update_param()}}{
Set estimated parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$update_param(val, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A vector with parameter values}

\item{\code{names}}{A vector with estimated parameter names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$runEstimation(
  maxiter,
  paramnames,
  metric,
  score_fn,
  weigh_fn,
  bounds,
  weathers,
  varID
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment}

\item{\code{bounds}}{A matrix with lower (col1) and upper (col2) bounds for}

\item{\code{weathers}}{A list of weather data for each environment}

\item{\code{varID}}{A numeric value of the index of variety currently estimating}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEvar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEvar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEvar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
