% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_weather.R
\name{TPEweather}
\alias{TPEweather}
\title{#' R6 Class Representing a TPE weather}
\description{
TODO
}
\details{
TODO
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of the weather}

\item{\code{wData}}{A dataframe with weather data}

\item{\code{wList}}{A list of weather dataframes to use when multiple weathers
are generated}

\item{\code{simuWeather}}{A dataframe with weather data for simulation}

\item{\code{dateParam}}{sowing date of grid point}

\item{\code{yearLevels}}{Years of generated weather data}

\item{\code{parent}}{Parent environment}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TPEweather-new}{\code{TPEweather$new()}}
\item \href{#method-TPEweather-set_weather}{\code{TPEweather$set_weather()}}
\item \href{#method-TPEweather-genClimate}{\code{TPEweather$genClimate()}}
\item \href{#method-TPEweather-set_dateParam}{\code{TPEweather$set_dateParam()}}
\item \href{#method-TPEweather-clone}{\code{TPEweather$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-new"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-new}{}}}
\subsection{Method \code{new()}}{
Create a new TPE weather object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$new(name = "0101", parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the TPE weather}

\item{\code{parent}}{Parent environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `TPEweather` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-set_weather"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-set_weather}{}}}
\subsection{Method \code{set_weather()}}{
Set weather data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$set_weather(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{A dataframe of weather data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-genClimate"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-genClimate}{}}}
\subsection{Method \code{genClimate()}}{
Generate climate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$genClimate(rcp, year, yearNb, modelNb, path, pathCLI, seed, bs = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rcp}}{Rcp scenario to use}

\item{\code{year}}{Year to simulate climate}

\item{\code{yearNb}}{Number of years to simulate}

\item{\code{modelNb}}{Identifier of model to use, see \code{generateClimate}}

\item{\code{path}}{Path to marksim standalone}

\item{\code{pathCLI}}{Optional. Path to CLI folder for marksim standalone}

\item{\code{seed}}{Integer number to use as seed for marksim weather generator}

\item{\code{bs}}{Boolean. Should weathers be stored for bootstrap}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-set_dateParam"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-set_dateParam}{}}}
\subsection{Method \code{set_dateParam()}}{
Set sowing date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$set_dateParam(year)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{year}}{Year of simulation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TPEweather-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TPEweather-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPEweather$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
