% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_estim.R
\name{VarEstim}
\alias{VarEstim}
\title{R6 Class representing a VarEstim Estimation process}
\description{
The `VarEstim` object contains all information to perform parameter
estimation for a given TPEaseVar variety
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string identifier of VarEstim}

\item{\code{parent}}{A TPEaseVar object parent of VarEstim}

\item{\code{environments}}{A list of TPEaseEnv objects to use for estimation}

\item{\code{parameters}}{A dataframe with all parameters (variety x environment)
used for simulation, each line is one environment}

\item{\code{DEparams}}{A DEoptim.control object with parameters used for DEoptim
parameter estimation}

\item{\code{DEresult}}{A DEoptim result object}

\item{\code{paramNames}}{A vector of estimated parameter names}

\item{\code{test}}{Debug}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VarEstim-new}{\code{VarEstim$new()}}
\item \href{#method-VarEstim-createEnv}{\code{VarEstim$createEnv()}}
\item \href{#method-VarEstim-runEstimation}{\code{VarEstim$runEstimation()}}
\item \href{#method-VarEstim-clone}{\code{VarEstim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-new"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-new}{}}}
\subsection{Method \code{new()}}{
Create a new VarEstim object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$new(
  name = "estim1",
  parent,
  environments,
  eparam,
  weathers,
  observations
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of VarEstim}

\item{\code{parent}}{A TPEaseVar object parent of VarEstim}

\item{\code{environments}}{A vector of environments names}

\item{\code{eparam}}{A dataframe of environments parameters}

\item{\code{weathers}}{A dataframe of environments weathers}

\item{\code{observations}}{A dataframe of environments observations for a given
TPEaseVar}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `VarEstim` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-createEnv"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-createEnv}{}}}
\subsection{Method \code{createEnv()}}{
Create a TPEaseEnv object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$createEnv(name, eparam, weather, observations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string identifier of the environment}

\item{\code{eparam}}{A dataframe of environment parameters}

\item{\code{weather}}{A dataframe of environment weather}

\item{\code{observations}}{A dataframe of environment observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-runEstimation"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-runEstimation}{}}}
\subsection{Method \code{runEstimation()}}{
Run parameter estimation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$runEstimation(
  maxiter = 2000,
  paramnames = NA,
  metric = "RMSE",
  score_fn = get_score,
  weigh_fn = NA,
  bounds = NA,
  args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxiter}}{A numeric value of the maximum number of iteration
for DEoptim}

\item{\code{paramnames}}{A vector of parameter names to be estimated}

\item{\code{metric}}{A character string with the name of the metric to use
for fitness computation. Options are c("RMSE","MAE","MSE")}

\item{\code{score_fn}}{A function to compute fitness, see \code{get_score}}

\item{\code{weigh_fn}}{Not used for the moment
TODO: implement weigh_fn}

\item{\code{bounds}}{A matrix with lower (col1) and upper (col2) bounds for
each of the parameters in `paramnames` (rows)}

\item{\code{args}}{Additional parameters to be passed to DEoptim.control}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VarEstim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VarEstim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VarEstim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
